refactor(rpc): controller service とテストを機能別ファイルに分割
proto/hdlctrl/v1/controller.proto のドメイン境界
(ホスト系/アカウント系/Cloud系/コンタクト・チャット系/セッション系)
に揃えて 6 ファイルずつに分割。ロジック変更なし。
