+++ adapter/rpc/controller_service.go
@@ -1,26 +1,18 @@
 package rpc
 
 import (
-	"context"
 	"errors"
-	"fmt"
 	"net/http"
-	"time"
 
 	"connectrpc.com/connect"
-	"github.com/hantabaru1014/baru-reso-headless-controller/adapter/converter"
-	"github.com/hantabaru1014/baru-reso-headless-controller/adapter/resonitelink"
 	"github.com/hantabaru1014/baru-reso-headless-controller/domain"
-	"github.com/hantabaru1014/baru-reso-headless-controller/domain/entity"
 	"github.com/hantabaru1014/baru-reso-headless-controller/lib/auth"
 	"github.com/hantabaru1014/baru-reso-headless-controller/lib/logging"
 	"github.com/hantabaru1014/baru-reso-headless-controller/lib/skyfrost"
 	hdlctrlv1 "github.com/hantabaru1014/baru-reso-headless-controller/pbgen/hdlctrl/v1"
 	"github.com/hantabaru1014/baru-reso-headless-controller/pbgen/hdlctrl/v1/hdlctrlv1connect"
-	headlessv1 "github.com/hantabaru1014/baru-reso-headless-controller/pbgen/headless/v1"
 	"github.com/hantabaru1014/baru-reso-headless-controller/usecase"
 	"github.com/hantabaru1014/baru-reso-headless-controller/usecase/port"
-	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 var _ hdlctrlv1connect.ControllerServiceHandler = (*ControllerService)(nil)
@@ -83,1099 +75,6 @@ func (c *ControllerService) NewHandler() (string, http.Handler) {
 	interceptors := connect.WithInterceptors(logging.NewErrorLogInterceptor(), auth.NewAuthInterceptor())
 
 	return hdlctrlv1connect.NewControllerServiceHandler(c, interceptors)
-}
-
-// ListHeadlessHostImageTags implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) ListHeadlessHostImageTags(ctx context.Context, req *connect.Request[hdlctrlv1.ListHeadlessHostImageTagsRequest]) (*connect.Response[hdlctrlv1.ListHeadlessHostImageTagsResponse], error) {
-	tags, err := c.hhrepo.ListContainerTags(ctx, nil)
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	protoTags := make([]*hdlctrlv1.ListHeadlessHostImageTagsResponse_ContainerImage, 0, len(tags))
-	for _, tag := range tags {
-		protoTags = append(protoTags, &hdlctrlv1.ListHeadlessHostImageTagsResponse_ContainerImage{
-			Tag:             tag.Tag,
-			ResoniteVersion: tag.ResoniteVersion,
-			IsPrerelease:    tag.IsPreRelease,
-			AppVersion:      tag.AppVersion,
-		})
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.ListHeadlessHostImageTagsResponse{
-		Tags: protoTags,
-	})
-
-	return res, nil
-}
-
-// StartHeadlessHost implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) StartHeadlessHost(ctx context.Context, req *connect.Request[hdlctrlv1.StartHeadlessHostRequest]) (*connect.Response[hdlctrlv1.StartHeadlessHostResponse], error) {
-	claims, err := auth.GetAuthClaimsFromContext(ctx)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeUnauthenticated, err)
-	}
-
-	account, err := c.hauc.GetHeadlessAccount(ctx, req.Msg.GetHeadlessAccountId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	params := port.HeadlessHostStartParams{
-		Name:              req.Msg.GetName(),
-		HeadlessAccount:   *account,
-		ContainerImageTag: req.Msg.GetImageTag(),
-		StartupConfig:     req.Msg.GetStartupConfig(),
-	}
-	if req.Msg.AutoUpdatePolicy != nil && req.Msg.GetAutoUpdatePolicy() != hdlctrlv1.HeadlessHostAutoUpdatePolicy_HEADLESS_HOST_AUTO_UPDATE_POLICY_UNKNOWN {
-		params.AutoUpdatePolicy = entity.HostAutoUpdatePolicy(req.Msg.GetAutoUpdatePolicy())
-	}
-
-	if req.Msg.Memo != nil {
-		params.Memo = req.Msg.GetMemo()
-	}
-
-	hostId, err := c.hhuc.HeadlessHostStart(ctx, params, &claims.UserID)
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.StartHeadlessHostResponse{
-		HostId: hostId,
-	})
-
-	return res, nil
-}
-
-// CreateHeadlessAccount implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) CreateHeadlessAccount(ctx context.Context, req *connect.Request[hdlctrlv1.CreateHeadlessAccountRequest]) (*connect.Response[hdlctrlv1.CreateHeadlessAccountResponse], error) {
-	err := c.hauc.CreateHeadlessAccount(ctx, req.Msg.GetCredential(), req.Msg.GetPassword())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.CreateHeadlessAccountResponse{})
-
-	return res, nil
-}
-
-// ListHeadlessAccounts implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) ListHeadlessAccounts(ctx context.Context, req *connect.Request[hdlctrlv1.ListHeadlessAccountsRequest]) (*connect.Response[hdlctrlv1.ListHeadlessAccountsResponse], error) {
-	pageIndex, pageSize, err := normalizePageRequest(req.Msg.GetPage())
-	if err != nil {
-		return nil, err
-	}
-
-	pageResult, err := c.hauc.ListHeadlessAccountsPaged(ctx, pageIndex, pageSize)
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	protoAccounts := make([]*hdlctrlv1.HeadlessAccount, 0, len(pageResult.Accounts))
-
-	for _, account := range pageResult.Accounts {
-		a := &hdlctrlv1.HeadlessAccount{
-			UserId: account.ResoniteID,
-		}
-		if account.LastDisplayName != nil {
-			a.UserName = *account.LastDisplayName
-		} else {
-			a.UserName = ""
-		}
-
-		if account.LastIconUrl != nil {
-			a.IconUrl = *account.LastIconUrl
-		} else {
-			a.IconUrl = ""
-		}
-
-		protoAccounts = append(protoAccounts, a)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.ListHeadlessAccountsResponse{
-		Accounts: protoAccounts,
-		Page: &hdlctrlv1.PageResponse{
-			TotalCount: pageResult.TotalCount,
-			PageIndex:  pageIndex,
-			PageSize:   pageSize,
-		},
-	})
-
-	return res, nil
-}
-
-// DeleteHeadlessAccount implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) DeleteHeadlessAccount(ctx context.Context, req *connect.Request[hdlctrlv1.DeleteHeadlessAccountRequest]) (*connect.Response[hdlctrlv1.DeleteHeadlessAccountResponse], error) {
-	err := c.hauc.DeleteHeadlessAccount(ctx, req.Msg.GetAccountId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	return connect.NewResponse(&hdlctrlv1.DeleteHeadlessAccountResponse{}), nil
-}
-
-// UpdateHeadlessAccountCredentials implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) UpdateHeadlessAccountCredentials(ctx context.Context, req *connect.Request[hdlctrlv1.UpdateHeadlessAccountCredentialsRequest]) (*connect.Response[hdlctrlv1.UpdateHeadlessAccountCredentialsResponse], error) {
-	err := c.hauc.UpdateHeadlessAccountCredentials(ctx, req.Msg.GetAccountId(), req.Msg.GetCredential(), req.Msg.GetPassword())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	return connect.NewResponse(&hdlctrlv1.UpdateHeadlessAccountCredentialsResponse{}), nil
-}
-
-// GetHeadlessAccountStorageInfo implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) GetHeadlessAccountStorageInfo(ctx context.Context, req *connect.Request[hdlctrlv1.GetHeadlessAccountStorageInfoRequest]) (*connect.Response[hdlctrlv1.GetHeadlessAccountStorageInfoResponse], error) {
-	account, err := c.hauc.GetHeadlessAccount(ctx, req.Msg.GetAccountId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	storageInfo, err := c.skyfrostClient.GetStorageInfo(ctx, account.Credential, account.Password, account.ResoniteID)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, fmt.Errorf("failed to get storage info for user: %w", err))
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.GetHeadlessAccountStorageInfoResponse{
-		StorageQuotaBytes: storageInfo.QuotaBytes,
-		StorageUsedBytes:  storageInfo.UsedBytes,
-	})
-
-	return res, nil
-}
-
-// RefetchHeadlessAccountInfo implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) RefetchHeadlessAccountInfo(ctx context.Context, req *connect.Request[hdlctrlv1.RefetchHeadlessAccountInfoRequest]) (*connect.Response[hdlctrlv1.RefetchHeadlessAccountInfoResponse], error) {
-	err := c.hauc.RefetchHeadlessAccountInfo(ctx, req.Msg.GetAccountId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	return connect.NewResponse(&hdlctrlv1.RefetchHeadlessAccountInfoResponse{}), nil
-}
-
-// UpdateHeadlessAccountIcon implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) UpdateHeadlessAccountIcon(ctx context.Context, req *connect.Request[hdlctrlv1.UpdateHeadlessAccountIconRequest]) (*connect.Response[hdlctrlv1.UpdateHeadlessAccountIconResponse], error) {
-	newIconUrl, err := c.hauc.UpdateHeadlessAccountIcon(ctx, req.Msg.GetAccountId(), req.Msg.GetIconData())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	return connect.NewResponse(&hdlctrlv1.UpdateHeadlessAccountIconResponse{
-		NewIconUrl: newIconUrl,
-	}), nil
-}
-
-// AcceptFriendRequests implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) AcceptFriendRequests(ctx context.Context, req *connect.Request[hdlctrlv1.AcceptFriendRequestsRequest]) (*connect.Response[hdlctrlv1.AcceptFriendRequestsResponse], error) {
-	hosts, err := c.hhrepo.ListRunningByAccount(ctx, req.Msg.GetHeadlessAccountId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	if len(hosts) == 0 {
-		return nil, connect.NewError(connect.CodeInvalidArgument, errors.New("このアカウントで起動中のヘッドレスホストが必要です"))
-	}
-
-	conn, err := c.hhrepo.GetRpcClient(ctx, hosts[0].ID)
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	ids := []string{
-		req.Msg.GetTargetUserId(),
-	}
-
-	_, err = conn.AcceptFriendRequests(ctx, &headlessv1.AcceptFriendRequestsRequest{
-		UserIds: ids,
-	})
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.AcceptFriendRequestsResponse{})
-
-	return res, nil
-}
-
-// GetFriendRequests implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) GetFriendRequests(ctx context.Context, req *connect.Request[hdlctrlv1.GetFriendRequestsRequest]) (*connect.Response[hdlctrlv1.GetFriendRequestsResponse], error) {
-	account, err := c.hauc.GetHeadlessAccount(ctx, req.Msg.GetHeadlessAccountId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	contacts, err := c.skyfrostClient.GetContacts(ctx, account.Credential, account.Password)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, fmt.Errorf("failed to get friend requests: %w", err))
-	}
-
-	requestedContacts := make([]*hdlctrlv1.UserInfo, 0)
-
-	for _, contact := range contacts {
-		if contact.Status == "Requested" {
-			requestedContacts = append(requestedContacts, &hdlctrlv1.UserInfo{
-				Id:      contact.Id,
-				Name:    contact.Username,
-				IconUrl: contact.Profile.IconUrl,
-			})
-		}
-	}
-
-	return connect.NewResponse(&hdlctrlv1.GetFriendRequestsResponse{
-		RequestedContacts: requestedContacts,
-	}), nil
-}
-
-// ListContacts implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) ListContacts(ctx context.Context, req *connect.Request[hdlctrlv1.ListContactsRequest]) (*connect.Response[hdlctrlv1.ListContactsResponse], error) {
-	hosts, err := c.hhrepo.ListRunningByAccount(ctx, req.Msg.GetHeadlessAccountId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	if len(hosts) == 0 {
-		return nil, connect.NewError(connect.CodeInvalidArgument, errors.New("このアカウントで起動中のヘッドレスホストが必要です"))
-	}
-
-	conn, err := c.hhrepo.GetRpcClient(ctx, hosts[0].ID)
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	headlessRes, err := conn.ListContacts(ctx, &headlessv1.ListContactsRequest{
-		Limit:  req.Msg.GetLimit(),
-		Cursor: req.Msg.Cursor,
-	})
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	contacts := make([]*hdlctrlv1.UserInfo, 0, len(headlessRes.GetUsers()))
-	for _, u := range headlessRes.GetUsers() {
-		contacts = append(contacts, &hdlctrlv1.UserInfo{
-			Id:      u.GetId(),
-			Name:    u.GetName(),
-			IconUrl: u.GetIconUrl(),
-		})
-	}
-
-	return connect.NewResponse(&hdlctrlv1.ListContactsResponse{
-		Contacts:   contacts,
-		NextCursor: headlessRes.NextCursor,
-	}), nil
-}
-
-// GetContactMessages implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) GetContactMessages(ctx context.Context, req *connect.Request[hdlctrlv1.GetContactMessagesRequest]) (*connect.Response[hdlctrlv1.GetContactMessagesResponse], error) {
-	account, err := c.hauc.GetHeadlessAccount(ctx, req.Msg.GetHeadlessAccountId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	hosts, err := c.hhrepo.ListRunningByAccount(ctx, req.Msg.GetHeadlessAccountId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	if len(hosts) == 0 {
-		return nil, connect.NewError(connect.CodeInvalidArgument, errors.New("このアカウントで起動中のヘッドレスホストが必要です"))
-	}
-
-	conn, err := c.hhrepo.GetRpcClient(ctx, hosts[0].ID)
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	headlessRes, err := conn.GetContactMessages(ctx, &headlessv1.GetContactMessagesRequest{
-		UserId:   req.Msg.GetContactUserId(),
-		Limit:    req.Msg.GetLimit(),
-		BeforeId: req.Msg.BeforeId,
-		AfterId:  req.Msg.AfterId,
-	})
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	messages := make([]*hdlctrlv1.ContactMessage, 0, len(headlessRes.GetMessages()))
-	for _, m := range headlessRes.GetMessages() {
-		messages = append(messages, &hdlctrlv1.ContactMessage{
-			Id:           m.GetId(),
-			Type:         m.GetType(),
-			Content:      m.GetContent(),
-			SendTime:     m.GetSendTime(),
-			ReadTime:     m.GetReadTime(),
-			IsOwnMessage: m.GetSenderId() == account.ResoniteID,
-		})
-	}
-
-	return connect.NewResponse(&hdlctrlv1.GetContactMessagesResponse{
-		Messages:      messages,
-		HasMoreBefore: headlessRes.GetHasMoreBefore(),
-		HasMoreAfter:  headlessRes.GetHasMoreAfter(),
-	}), nil
-}
-
-// SendContactMessage implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) SendContactMessage(ctx context.Context, req *connect.Request[hdlctrlv1.SendContactMessageRequest]) (*connect.Response[hdlctrlv1.SendContactMessageResponse], error) {
-	hosts, err := c.hhrepo.ListRunningByAccount(ctx, req.Msg.GetHeadlessAccountId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	if len(hosts) == 0 {
-		return nil, connect.NewError(connect.CodeInvalidArgument, errors.New("このアカウントで起動中のヘッドレスホストが必要です"))
-	}
-
-	conn, err := c.hhrepo.GetRpcClient(ctx, hosts[0].ID)
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	_, err = conn.SendContactMessage(ctx, &headlessv1.SendContactMessageRequest{
-		UserId:  req.Msg.GetContactUserId(),
-		Message: req.Msg.GetMessage(),
-	})
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	return connect.NewResponse(&hdlctrlv1.SendContactMessageResponse{}), nil
-}
-
-// RestartHeadlessHost implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) RestartHeadlessHost(ctx context.Context, req *connect.Request[hdlctrlv1.RestartHeadlessHostRequest]) (*connect.Response[hdlctrlv1.RestartHeadlessHostResponse], error) {
-	var newTag *string
-
-	if req.Msg.GetWithUpdate() {
-		str := "latestRelease"
-		newTag = &str
-	} else if req.Msg.GetWithImageTag() != "" {
-		newTag = req.Msg.WithImageTag
-	}
-
-	timeout := defaultRestartTimeoutSeconds
-	if req.Msg.TimeoutSeconds != nil {
-		timeout = int(req.Msg.GetTimeoutSeconds())
-	}
-
-	err := c.hhuc.HeadlessHostRestart(ctx, req.Msg.GetHostId(), newTag, req.Msg.GetWithWorldRestart(), timeout)
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.RestartHeadlessHostResponse{
-		NewHostId: &req.Msg.HostId,
-	})
-
-	return res, nil
-}
-
-// UpdateHeadlessHostSettings implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) UpdateHeadlessHostSettings(ctx context.Context, req *connect.Request[hdlctrlv1.UpdateHeadlessHostSettingsRequest]) (*connect.Response[hdlctrlv1.UpdateHeadlessHostSettingsResponse], error) {
-	host, err := c.hhuc.HeadlessHostGet(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	if req.Msg.Name != nil {
-		err := c.hhrepo.Rename(ctx, req.Msg.GetHostId(), req.Msg.GetName())
-		if err != nil {
-			return nil, convertErr(err)
-		}
-	}
-
-	if req.Msg.AutoUpdatePolicy != nil &&
-		req.Msg.GetAutoUpdatePolicy() != hdlctrlv1.HeadlessHostAutoUpdatePolicy_HEADLESS_HOST_AUTO_UPDATE_POLICY_UNKNOWN {
-		err := c.hhrepo.UpdateAutoUpdatePolicy(
-			ctx,
-			req.Msg.GetHostId(),
-			entity.HostAutoUpdatePolicy(req.Msg.GetAutoUpdatePolicy()),
-		)
-		if err != nil {
-			return nil, convertErr(err)
-		}
-	}
-
-	hasUpdateReq := false
-	updateReq := &headlessv1.UpdateHostSettingsRequest{}
-	settings := host.HostSettings
-
-	if req.Msg.TickRate != nil {
-		hasUpdateReq = true
-		tickRate := req.Msg.GetTickRate()
-		updateReq.TickRate = &tickRate
-		settings.TickRate = tickRate
-	}
-
-	if req.Msg.MaxConcurrentAssetTransfers != nil {
-		hasUpdateReq = true
-		maxConcurrentAssetTransfers := req.Msg.GetMaxConcurrentAssetTransfers()
-		updateReq.MaxConcurrentAssetTransfers = &maxConcurrentAssetTransfers
-		settings.MaxConcurrentAssetTransfers = maxConcurrentAssetTransfers
-	}
-
-	if req.Msg.UsernameOverride != nil {
-		hasUpdateReq = true
-		usernameOverride := req.Msg.GetUsernameOverride()
-		updateReq.UsernameOverride = &usernameOverride
-		settings.UsernameOverride = &usernameOverride
-	}
-
-	if req.Msg.GetUpdateAutoSpawnItems() {
-		hasUpdateReq = true
-		updateReq.UpdateAutoSpawnItems = true
-		updateReq.AutoSpawnItems = req.Msg.GetAutoSpawnItems()
-		settings.AutoSpawnItems = req.Msg.GetAutoSpawnItems()
-	}
-
-	if req.Msg.UniverseId != nil {
-		hasUpdateReq = true
-		// 実行中のホストのUniverseIDの更新は未対応
-		settings.UniverseID = req.Msg.UniverseId
-	}
-
-	if hasUpdateReq {
-		if host.Status == entity.HeadlessHostStatus_RUNNING {
-			conn, err := c.hhrepo.GetRpcClient(ctx, req.Msg.GetHostId())
-			if err != nil {
-				return nil, convertRpcClientErr(err)
-			}
-
-			_, err = conn.UpdateHostSettings(ctx, updateReq)
-			if err != nil {
-				return nil, convertRpcClientErr(err)
-			}
-
-			updated, err := conn.GetStartupConfigToRestore(ctx, &headlessv1.GetStartupConfigToRestoreRequest{
-				IncludeStartWorlds: false,
-			})
-			if err != nil {
-				return nil, convertRpcClientErr(err)
-			}
-
-			err = c.hhrepo.UpdateHostSettings(ctx, req.Msg.GetHostId(), converter.HeadlessHostSettingsProtoToEntity(updated.GetStartupConfig()))
-			if err != nil {
-				return nil, convertErr(err)
-			}
-		} else {
-			err = c.hhrepo.UpdateHostSettings(ctx, req.Msg.GetHostId(), &settings)
-			if err != nil {
-				return nil, convertErr(err)
-			}
-		}
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.UpdateHeadlessHostSettingsResponse{})
-
-	return res, nil
-}
-
-// GetHeadlessHostLogs implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) GetHeadlessHostLogs(ctx context.Context, req *connect.Request[hdlctrlv1.GetHeadlessHostLogsRequest]) (*connect.Response[hdlctrlv1.GetHeadlessHostLogsResponse], error) {
-	// カーソル解析 (ID-based)
-	var beforeID, afterID int64
-
-	switch cursor := req.Msg.GetCursor().(type) {
-	case *hdlctrlv1.GetHeadlessHostLogsRequest_BeforeId:
-		beforeID = cursor.BeforeId
-	case *hdlctrlv1.GetHeadlessHostLogsRequest_AfterId:
-		afterID = cursor.AfterId
-	}
-
-	result, err := c.hhuc.HeadlessHostGetLogs(ctx, usecase.HeadlessHostGetLogsParams{
-		HostID:     req.Msg.GetHostId(),
-		InstanceID: req.Msg.GetInstanceId(),
-		Limit:      req.Msg.GetLimit(),
-		BeforeID:   beforeID,
-		AfterID:    afterID,
-	})
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	protoLogs := make([]*hdlctrlv1.GetHeadlessHostLogsResponse_Log, 0, len(result.Logs))
-	for _, log := range result.Logs {
-		protoLogs = append(protoLogs, &hdlctrlv1.GetHeadlessHostLogsResponse_Log{
-			Timestamp: timestamppb.New(time.Unix(log.Timestamp, 0)),
-			IsError:   log.IsError,
-			Body:      log.Body,
-			Id:        log.ID,
-		})
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.GetHeadlessHostLogsResponse{
-		Logs:          protoLogs,
-		HasMoreBefore: result.HasMoreBefore,
-		HasMoreAfter:  result.HasMoreAfter,
-	})
-
-	return res, nil
-}
-
-// ListHeadlessHostInstances implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) ListHeadlessHostInstances(ctx context.Context, req *connect.Request[hdlctrlv1.ListHeadlessHostInstancesRequest]) (*connect.Response[hdlctrlv1.ListHeadlessHostInstancesResponse], error) {
-	instances, err := c.hhuc.HeadlessHostGetInstances(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	protoInstances := make([]*hdlctrlv1.ListHeadlessHostInstancesResponse_Instance, 0, len(instances))
-
-	for _, inst := range instances {
-		protoInst := &hdlctrlv1.ListHeadlessHostInstancesResponse_Instance{
-			InstanceId: inst.InstanceID,
-			LogCount:   inst.LogCount,
-			IsCurrent:  inst.IsCurrent,
-		}
-		if inst.FirstLogAt != nil {
-			protoInst.FirstLogAt = timestamppb.New(time.Unix(*inst.FirstLogAt, 0))
-		}
-
-		if inst.LastLogAt != nil {
-			protoInst.LastLogAt = timestamppb.New(time.Unix(*inst.LastLogAt, 0))
-		}
-
-		protoInstances = append(protoInstances, protoInst)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.ListHeadlessHostInstancesResponse{
-		Instances: protoInstances,
-	})
-
-	return res, nil
-}
-
-// ShutdownHeadlessHost implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) ShutdownHeadlessHost(ctx context.Context, req *connect.Request[hdlctrlv1.ShutdownHeadlessHostRequest]) (*connect.Response[hdlctrlv1.ShutdownHeadlessHostResponse], error) {
-	err := c.hhuc.HeadlessHostShutdown(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.ShutdownHeadlessHostResponse{})
-
-	return res, nil
-}
-
-// KillHeadlessHost implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) KillHeadlessHost(ctx context.Context, req *connect.Request[hdlctrlv1.KillHeadlessHostRequest]) (*connect.Response[hdlctrlv1.KillHeadlessHostResponse], error) {
-	err := c.hhuc.HeadlessHostKill(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.KillHeadlessHostResponse{})
-
-	return res, nil
-}
-
-// AllowHostAccess implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) AllowHostAccess(ctx context.Context, req *connect.Request[hdlctrlv1.AllowHostAccessRequest]) (*connect.Response[hdlctrlv1.AllowHostAccessResponse], error) {
-	conn, err := c.hhrepo.GetRpcClient(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	_, err = conn.AllowHostAccess(ctx, req.Msg.GetRequest())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.AllowHostAccessResponse{})
-
-	return res, nil
-}
-
-// DenyHostAccess implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) DenyHostAccess(ctx context.Context, req *connect.Request[hdlctrlv1.DenyHostAccessRequest]) (*connect.Response[hdlctrlv1.DenyHostAccessResponse], error) {
-	conn, err := c.hhrepo.GetRpcClient(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	_, err = conn.DenyHostAccess(ctx, req.Msg.GetRequest())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.DenyHostAccessResponse{})
-
-	return res, nil
-}
-
-// BanUser implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) BanUser(ctx context.Context, req *connect.Request[hdlctrlv1.BanUserRequest]) (*connect.Response[hdlctrlv1.BanUserResponse], error) {
-	conn, err := c.hhrepo.GetRpcClient(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	_, err = conn.BanUser(ctx, req.Msg.GetParameters())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.BanUserResponse{})
-
-	return res, nil
-}
-
-// KickUser implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) KickUser(ctx context.Context, req *connect.Request[hdlctrlv1.KickUserRequest]) (*connect.Response[hdlctrlv1.KickUserResponse], error) {
-	conn, err := c.hhrepo.GetRpcClient(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	_, err = conn.KickUser(ctx, req.Msg.GetParameters())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.KickUserResponse{})
-
-	return res, nil
-}
-
-// SearchUserInfo implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) SearchUserInfo(ctx context.Context, req *connect.Request[hdlctrlv1.SearchUserInfoRequest]) (*connect.Response[headlessv1.SearchUserInfoResponse], error) {
-	conn, err := c.hhrepo.GetRpcClient(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	headlessRes, err := conn.SearchUserInfo(ctx, req.Msg.GetParameters())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	res := connect.NewResponse(headlessRes)
-
-	return res, nil
-}
-
-// FetchWorldInfo implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) FetchWorldInfo(ctx context.Context, req *connect.Request[hdlctrlv1.FetchWorldInfoRequest]) (*connect.Response[headlessv1.FetchWorldInfoResponse], error) {
-	conn, err := c.hhrepo.GetRpcClient(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	headlessRes, err := conn.FetchWorldInfo(ctx, &headlessv1.FetchWorldInfoRequest{
-		Url: req.Msg.GetUrl(),
-	})
-	if err != nil {
-		return nil, connect.NewError(connect.CodeNotFound, err)
-	}
-
-	res := connect.NewResponse(headlessRes)
-
-	return res, nil
-}
-
-// GetResoniteUser implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) GetResoniteUser(ctx context.Context, req *connect.Request[hdlctrlv1.GetResoniteUserRequest]) (*connect.Response[hdlctrlv1.GetResoniteUserResponse], error) {
-	userInfo, err := c.skyfrostClient.FetchUserInfo(ctx, req.Msg.GetResoniteId())
-	if err != nil {
-		return nil, connect.NewError(connect.CodeNotFound, err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.GetResoniteUserResponse{
-		Id:      userInfo.ID,
-		Name:    userInfo.UserName,
-		IconUrl: userInfo.IconUrl,
-	})
-
-	return res, nil
-}
-
-// GetOwnWorlds implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) GetOwnWorlds(ctx context.Context, req *connect.Request[hdlctrlv1.GetOwnWorldsRequest]) (*connect.Response[hdlctrlv1.GetOwnWorldsResponse], error) {
-	host, err := c.hhuc.HeadlessHostGet(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	account, err := c.hauc.GetHeadlessAccount(ctx, host.AccountId)
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	result, err := c.skyfrostClient.GetOwnWorlds(ctx, account.Credential, account.Password, int(req.Msg.GetPageIndex()))
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, fmt.Errorf("failed to get own worlds: %w", err))
-	}
-
-	records := make([]*hdlctrlv1.SearchWorldsResponse_WorldRecord, 0, len(result.Records))
-	for _, r := range result.Records {
-		records = append(records, &hdlctrlv1.SearchWorldsResponse_WorldRecord{
-			Id:           r.ID,
-			OwnerId:      r.OwnerID,
-			OwnerName:    r.OwnerName,
-			Name:         r.Name,
-			Description:  r.Description,
-			ThumbnailUrl: r.ThumbnailUri,
-			IsFeatured:   r.IsFeatured,
-		})
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.GetOwnWorldsResponse{
-		Records: records,
-		HasMore: result.HasMore,
-	})
-
-	return res, nil
-}
-
-// SearchWorlds implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) SearchWorlds(ctx context.Context, req *connect.Request[hdlctrlv1.SearchWorldsRequest]) (*connect.Response[hdlctrlv1.SearchWorldsResponse], error) {
-	result, err := c.skyfrostClient.SearchWorlds(ctx, req.Msg.GetQuery(), req.Msg.GetFeaturedOnly(), int(req.Msg.GetPageIndex()))
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, err)
-	}
-
-	records := make([]*hdlctrlv1.SearchWorldsResponse_WorldRecord, 0, len(result.Records))
-	for _, r := range result.Records {
-		records = append(records, &hdlctrlv1.SearchWorldsResponse_WorldRecord{
-			Id:           r.ID,
-			OwnerId:      r.OwnerID,
-			OwnerName:    r.OwnerName,
-			Name:         r.Name,
-			Description:  r.Description,
-			ThumbnailUrl: r.ThumbnailUri,
-			IsFeatured:   r.IsFeatured,
-		})
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.SearchWorldsResponse{
-		Records: records,
-		HasMore: result.HasMore,
-	})
-
-	return res, nil
-}
-
-// GetHeadlessHost implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) GetHeadlessHost(ctx context.Context, req *connect.Request[hdlctrlv1.GetHeadlessHostRequest]) (*connect.Response[hdlctrlv1.GetHeadlessHostResponse], error) {
-	host, err := c.hhuc.HeadlessHostGet(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.GetHeadlessHostResponse{
-		Host: converter.HeadlessHostEntityToProto(host),
-	})
-
-	return res, nil
-}
-
-// ListHeadlessHost implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) ListHeadlessHost(ctx context.Context, req *connect.Request[hdlctrlv1.ListHeadlessHostRequest]) (*connect.Response[hdlctrlv1.ListHeadlessHostResponse], error) {
-	pageIndex, pageSize, err := normalizePageRequest(req.Msg.GetPage())
-	if err != nil {
-		return nil, err
-	}
-
-	pageResult, err := c.hhuc.HeadlessHostListPaged(ctx, pageIndex, pageSize)
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	protoHosts := make([]*hdlctrlv1.HeadlessHost, 0, len(pageResult.Hosts))
-	for _, host := range pageResult.Hosts {
-		protoHosts = append(protoHosts, converter.HeadlessHostEntityToProto(host))
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.ListHeadlessHostResponse{
-		Hosts: protoHosts,
-		Page: &hdlctrlv1.PageResponse{
-			TotalCount: pageResult.TotalCount,
-			PageIndex:  pageIndex,
-			PageSize:   pageSize,
-		},
-	})
-
-	return res, nil
-}
-
-// DeleteHeadlessHost implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) DeleteHeadlessHost(ctx context.Context, req *connect.Request[hdlctrlv1.DeleteHeadlessHostRequest]) (*connect.Response[hdlctrlv1.DeleteHeadlessHostResponse], error) {
-	err := c.hhuc.HeadlessHostDelete(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	return connect.NewResponse(&hdlctrlv1.DeleteHeadlessHostResponse{}), nil
-}
-
-// GetSessionDetails implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) GetSessionDetails(ctx context.Context, req *connect.Request[hdlctrlv1.GetSessionDetailsRequest]) (*connect.Response[hdlctrlv1.GetSessionDetailsResponse], error) {
-	s, err := c.suc.GetSession(ctx, req.Msg.GetSessionId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.GetSessionDetailsResponse{
-		Session: converter.SessionEntityToProto(s),
-	})
-
-	return res, nil
-}
-
-// ListUsersInSession implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) ListUsersInSession(ctx context.Context, req *connect.Request[hdlctrlv1.ListUsersInSessionRequest]) (*connect.Response[hdlctrlv1.ListUsersInSessionResponse], error) {
-	conn, err := c.hhrepo.GetRpcClient(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	headlessRes, err := conn.ListUsersInSession(ctx, &headlessv1.ListUsersInSessionRequest{SessionId: req.Msg.GetSessionId()})
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.ListUsersInSessionResponse{
-		Users: headlessRes.GetUsers(),
-	})
-
-	return res, nil
-}
-
-// SaveSessionWorld implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) SaveSessionWorld(ctx context.Context, req *connect.Request[hdlctrlv1.SaveSessionWorldRequest]) (*connect.Response[hdlctrlv1.SaveSessionWorldResponse], error) {
-	var saveMode usecase.SaveMode
-
-	switch req.Msg.GetSaveMode() {
-	case hdlctrlv1.SaveSessionWorldRequest_SAVE_MODE_OVERWRITE:
-		saveMode = usecase.SaveMode_OVERWRITE
-	case hdlctrlv1.SaveSessionWorldRequest_SAVE_MODE_SAVE_AS:
-		saveMode = usecase.SaveMode_SAVE_AS
-	case hdlctrlv1.SaveSessionWorldRequest_SAVE_MODE_COPY:
-		saveMode = usecase.SaveMode_COPY
-	case hdlctrlv1.SaveSessionWorldRequest_SAVE_MODE_UNKNOWN:
-		return nil, connect.NewError(connect.CodeInvalidArgument, fmt.Errorf("invalid save mode: %s", req.Msg.GetSaveMode().String()))
-	}
-
-	savedRecordUrl, err := c.suc.SaveSessionWorld(ctx, req.Msg.GetSessionId(), saveMode)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInternal, err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.SaveSessionWorldResponse{
-		SavedRecordUrl: &savedRecordUrl,
-	})
-
-	return res, nil
-}
-
-// PrepareSessionWorldDownload implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) PrepareSessionWorldDownload(ctx context.Context, req *connect.Request[hdlctrlv1.PrepareSessionWorldDownloadRequest]) (*connect.Response[hdlctrlv1.PrepareSessionWorldDownloadResponse], error) {
-	if req.Msg.GetFormat() == headlessv1.WorldBinaryFormat_WORLD_BINARY_FORMAT_UNSPECIFIED {
-		return nil, connect.NewError(connect.CodeInvalidArgument, errors.New("format is required"))
-	}
-
-	url, filename, err := c.buc.PrepareSessionWorldDownload(ctx, req.Msg.GetSessionId(), req.Msg.GetFormat())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	return connect.NewResponse(&hdlctrlv1.PrepareSessionWorldDownloadResponse{
-		DownloadUrl: url,
-		Filename:    filename,
-	}), nil
-}
-
-// UpdateSessionParameters implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) UpdateSessionParameters(ctx context.Context, req *connect.Request[hdlctrlv1.UpdateSessionParametersRequest]) (*connect.Response[hdlctrlv1.UpdateSessionParametersResponse], error) {
-	err := c.suc.UpdateSessionParameters(ctx, req.Msg.GetParameters().GetSessionId(), req.Msg.GetParameters())
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInvalidArgument, err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.UpdateSessionParametersResponse{})
-
-	return res, nil
-}
-
-// UpdateUserRole implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) UpdateUserRole(ctx context.Context, req *connect.Request[hdlctrlv1.UpdateUserRoleRequest]) (*connect.Response[hdlctrlv1.UpdateUserRoleResponse], error) {
-	conn, err := c.hhrepo.GetRpcClient(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	headlessRes, err := conn.UpdateUserRole(ctx, req.Msg.GetParameters())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.UpdateUserRoleResponse{
-		Role: headlessRes.GetRole(),
-	})
-
-	return res, nil
-}
-
-// StartWorld implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) StartWorld(ctx context.Context, req *connect.Request[hdlctrlv1.StartWorldRequest]) (*connect.Response[hdlctrlv1.StartWorldResponse], error) {
-	claims, err := auth.GetAuthClaimsFromContext(ctx)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeUnauthenticated, err)
-	}
-
-	openedSession, err := c.suc.StartSession(ctx, req.Msg.GetHostId(), &claims.UserID, req.Msg.GetParameters(), &req.Msg.Memo)
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.StartWorldResponse{
-		OpenedSession: converter.SessionEntityToProto(openedSession),
-	})
-
-	return res, nil
-}
-
-// InviteUser implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) InviteUser(ctx context.Context, req *connect.Request[hdlctrlv1.InviteUserRequest]) (*connect.Response[hdlctrlv1.InviteUserResponse], error) {
-	conn, err := c.hhrepo.GetRpcClient(ctx, req.Msg.GetHostId())
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	hreq := &headlessv1.InviteUserRequest{
-		SessionId: req.Msg.GetSessionId(),
-	}
-	if req.Msg.GetUserId() != "" {
-		hreq.User = &headlessv1.InviteUserRequest_UserId{UserId: req.Msg.GetUserId()}
-	} else {
-		hreq.User = &headlessv1.InviteUserRequest_UserName{UserName: req.Msg.GetUserName()}
-	}
-
-	_, err = conn.InviteUser(ctx, hreq)
-	if err != nil {
-		return nil, convertRpcClientErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.InviteUserResponse{})
-
-	return res, nil
-}
-
-// IssueResoniteLinkConnection implements hdlctrlv1connect.ControllerServiceHandler.
-// 認証済みユーザに対し ResoniteLink WebSocket 接続用の path?query を返す。
-// 返される ws_path は host を含まない相対パスで、クライアントが現在の origin を補完して使う。
-func (c *ControllerService) IssueResoniteLinkConnection(ctx context.Context, req *connect.Request[hdlctrlv1.IssueResoniteLinkConnectionRequest]) (*connect.Response[hdlctrlv1.IssueResoniteLinkConnectionResponse], error) {
-	claims, err := auth.GetAuthClaimsFromContext(ctx)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeUnauthenticated, err)
-	}
-
-	if req.Msg.GetSessionId() == "" {
-		return nil, connect.NewError(connect.CodeInvalidArgument, errors.New("session_id is required"))
-	}
-
-	token, expiresAt, err := c.suc.IssueResoniteLinkToken(ctx, req.Msg.GetSessionId(), claims.UserID)
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.IssueResoniteLinkConnectionResponse{
-		WsPath:    resonitelink.BuildWSPath(token),
-		ExpiresAt: timestamppb.New(expiresAt),
-	})
-
-	return res, nil
-}
-
-// StopSession implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) StopSession(ctx context.Context, req *connect.Request[hdlctrlv1.StopSessionRequest]) (*connect.Response[hdlctrlv1.StopSessionResponse], error) {
-	err := c.suc.StopSession(ctx, req.Msg.GetSessionId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.StopSessionResponse{})
-
-	return res, nil
-}
-
-// SearchSessions implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) SearchSessions(ctx context.Context, req *connect.Request[hdlctrlv1.SearchSessionsRequest]) (*connect.Response[hdlctrlv1.SearchSessionsResponse], error) {
-	pageIndex, pageSize, err := normalizePageRequest(req.Msg.GetPage())
-	if err != nil {
-		return nil, err
-	}
-
-	filter := usecase.SearchSessionsFilter{
-		HostID:    req.Msg.GetParameters().HostId,
-		PageIndex: pageIndex,
-		PageSize:  pageSize,
-	}
-
-	if req.Msg.GetParameters() != nil && req.Msg.GetParameters().Status != nil {
-		s := entity.SessionStatus(int32(req.Msg.GetParameters().GetStatus().Number()))
-		filter.Status = &s
-	}
-
-	result, err := c.suc.SearchSessions(ctx, filter)
-	if err != nil {
-		return nil, connect.NewError(connect.CodeInvalidArgument, err)
-	}
-
-	protoSessions := make([]*hdlctrlv1.Session, 0, len(result.Sessions))
-	for _, session := range result.Sessions {
-		protoSessions = append(protoSessions, converter.SessionEntityToProto(session))
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.SearchSessionsResponse{
-		Sessions: protoSessions,
-		Page: &hdlctrlv1.PageResponse{
-			TotalCount: result.TotalCount,
-			PageIndex:  pageIndex,
-			PageSize:   pageSize,
-		},
-	})
-
-	return res, nil
-}
-
-// DeleteEndedSession implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) DeleteEndedSession(ctx context.Context, req *connect.Request[hdlctrlv1.DeleteEndedSessionRequest]) (*connect.Response[hdlctrlv1.DeleteEndedSessionResponse], error) {
-	err := c.suc.DeleteSession(ctx, req.Msg.GetSessionId())
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.DeleteEndedSessionResponse{})
-
-	return res, nil
-}
-
-// UpdateSessionExtraSettings implements hdlctrlv1connect.ControllerServiceHandler.
-func (c *ControllerService) UpdateSessionExtraSettings(ctx context.Context, req *connect.Request[hdlctrlv1.UpdateSessionExtraSettingsRequest]) (*connect.Response[hdlctrlv1.UpdateSessionExtraSettingsResponse], error) {
-	// TODO: いい感じにusecaseに移動する
-	s, err := c.suc.GetSession(ctx, req.Msg.GetSessionId())
-	if err != nil {
-		return nil, connect.NewError(connect.CodeNotFound, err)
-	}
-
-	if req.Msg.AutoUpgrade != nil {
-		s.AutoUpgrade = req.Msg.GetAutoUpgrade()
-	}
-
-	if req.Msg.Memo != nil {
-		s.Memo = req.Msg.GetMemo()
-	}
-
-	err = c.srepo.Upsert(ctx, s)
-	if err != nil {
-		return nil, convertErr(err)
-	}
-
-	res := connect.NewResponse(&hdlctrlv1.UpdateSessionExtraSettingsResponse{})
-
-	return res, nil
 }
 
 // convertErr converts domain errors to appropriate Connect RPC error codes.
