+++ adapter/rpc/controller_service_test.go
@@ -1,33 +1,22 @@
 package rpc
 
 import (
-	"context"
-	"errors"
-	"fmt"
-	"net/url"
-	"strings"
 	"testing"
-	"time"
 
 	"connectrpc.com/connect"
 	"github.com/hantabaru1014/baru-reso-headless-controller/adapter"
-	"github.com/hantabaru1014/baru-reso-headless-controller/adapter/hostconnector"
 	hostconnectormock "github.com/hantabaru1014/baru-reso-headless-controller/adapter/hostconnector/mock"
 	"github.com/hantabaru1014/baru-reso-headless-controller/adapter/sessionstate"
 	"github.com/hantabaru1014/baru-reso-headless-controller/config"
 	"github.com/hantabaru1014/baru-reso-headless-controller/db"
-	"github.com/hantabaru1014/baru-reso-headless-controller/domain/entity"
 	"github.com/hantabaru1014/baru-reso-headless-controller/lib/auth"
 	blobstoremock "github.com/hantabaru1014/baru-reso-headless-controller/lib/blobstore/mock"
-	"github.com/hantabaru1014/baru-reso-headless-controller/lib/skyfrost"
 	skyfrostmock "github.com/hantabaru1014/baru-reso-headless-controller/lib/skyfrost/mock"
 	hdlctrlv1 "github.com/hantabaru1014/baru-reso-headless-controller/pbgen/hdlctrl/v1"
 	"github.com/hantabaru1014/baru-reso-headless-controller/pbgen/hdlctrl/v1/hdlctrlv1connect"
-	headlessv1 "github.com/hantabaru1014/baru-reso-headless-controller/pbgen/headless/v1"
 	"github.com/hantabaru1014/baru-reso-headless-controller/testutil"
 	"github.com/hantabaru1014/baru-reso-headless-controller/usecase"
 	"github.com/hantabaru1014/baru-reso-headless-controller/usecase/port"
-	"github.com/jackc/pgx/v5/pgtype"
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -108,271 +97,6 @@ func setupAuthenticatedClient(t *testing.T, service *ControllerService) hdlctrlv
 		server.Client(),
 		server.URL,
 	)
-}
-
-func TestControllerService_ListHeadlessAccounts(t *testing.T) {
-	setup := setupControllerServiceTest(t)
-	defer setup.Cleanup()
-
-	client := setupAuthenticatedClient(t, setup.service)
-
-	// Create 25 test headless accounts for pagination
-	const totalAccounts = 25
-	for i := 1; i <= totalAccounts; i++ {
-		id := fmt.Sprintf("U-test%02d", i)
-		testutil.CreateTestHeadlessAccount(t, setup.queries, id, id+"@example.test", "password")
-	}
-
-	t.Run("成功: ヘッドレスアカウントのリストを取得 (ページング検証)", func(t *testing.T) {
-		// page 未指定 -> デフォルト 20 件
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListHeadlessAccountsRequest{})
-
-		res, err := client.ListHeadlessAccounts(t.Context(), req)
-		require.NoError(t, err)
-
-		assert.Len(t, res.Msg.GetAccounts(), 20)
-		require.NotNil(t, res.Msg.GetPage())
-		assert.Equal(t, int32(totalAccounts), res.Msg.GetPage().GetTotalCount())
-		assert.Equal(t, int32(0), res.Msg.GetPage().GetPageIndex())
-		assert.Equal(t, int32(20), res.Msg.GetPage().GetPageSize())
-
-		// Verify account data
-		account1 := res.Msg.GetAccounts()[0]
-		assert.NotEmpty(t, account1.GetUserId())
-		assert.NotEmpty(t, account1.GetUserName())
-
-		// page_index=1, page_size=20 -> 残り 5 件
-		req2 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListHeadlessAccountsRequest{
-			Page: &hdlctrlv1.PageRequest{PageIndex: 1, PageSize: 20},
-		})
-		res2, err := client.ListHeadlessAccounts(t.Context(), req2)
-		require.NoError(t, err)
-		assert.Len(t, res2.Msg.GetAccounts(), 5)
-		assert.Equal(t, int32(totalAccounts), res2.Msg.GetPage().GetTotalCount())
-
-		// page_size=50 -> 全件
-		req3 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListHeadlessAccountsRequest{
-			Page: &hdlctrlv1.PageRequest{PageIndex: 0, PageSize: 50},
-		})
-		res3, err := client.ListHeadlessAccounts(t.Context(), req3)
-		require.NoError(t, err)
-		assert.Len(t, res3.Msg.GetAccounts(), totalAccounts)
-
-		// page_size=150 -> 100 にクランプ。25件しかないので全件返るが PageSize は 100 として返る
-		req4 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListHeadlessAccountsRequest{
-			Page: &hdlctrlv1.PageRequest{PageIndex: 0, PageSize: 150},
-		})
-		res4, err := client.ListHeadlessAccounts(t.Context(), req4)
-		require.NoError(t, err)
-		assert.Len(t, res4.Msg.GetAccounts(), totalAccounts)
-		assert.Equal(t, int32(100), res4.Msg.GetPage().GetPageSize())
-	})
-}
-
-func TestControllerService_CreateHeadlessAccount(t *testing.T) {
-	t.Run("成功: 有効な認証情報でアカウント作成", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Mock skyfrost client to return successful login
-		setup.mockSkyfrost.EXPECT().
-			UserLogin(gomock.Any(), "testuser@example.test", "testpass123").
-			Return(&skyfrost.UserSession{UserId: "U-testuser123"}, nil)
-
-		setup.mockSkyfrost.EXPECT().
-			FetchUserInfo(gomock.Any(), "U-testuser123").
-			Return(&skyfrost.UserInfo{
-				ID:       "U-testuser123",
-				UserName: "TestUser",
-				IconUrl:  "https://example.test/icon.png",
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.CreateHeadlessAccountRequest{
-			Credential: "testuser@example.test",
-			Password:   "testpass123",
-		})
-
-		res, err := client.CreateHeadlessAccount(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-
-		// Verify account was created in DB
-		account, err := setup.queries.GetHeadlessAccount(t.Context(), "U-testuser123")
-		require.NoError(t, err)
-		assert.Equal(t, "testuser@example.test", account.Credential)
-	})
-
-	t.Run("失敗: 無効な認証情報でアカウント作成", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Mock skyfrost client to return login error
-		setup.mockSkyfrost.EXPECT().
-			UserLogin(gomock.Any(), "invalid@example.test", "invalidpassword").
-			Return(nil, connect.NewError(connect.CodeUnauthenticated, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.CreateHeadlessAccountRequest{
-			Credential: "invalid@example.test",
-			Password:   "invalidpassword",
-		})
-
-		_, err := client.CreateHeadlessAccount(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-
-	t.Run("失敗: 既に存在するアカウントを作成", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create initial account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-existing", "existing@example.test", "password123")
-
-		// Mock skyfrost to return successful login but DB insert will fail
-		setup.mockSkyfrost.EXPECT().
-			UserLogin(gomock.Any(), "existing@example.test", "newpassword123").
-			Return(&skyfrost.UserSession{UserId: "U-existing"}, nil)
-
-		setup.mockSkyfrost.EXPECT().
-			FetchUserInfo(gomock.Any(), "U-existing").
-			Return(&skyfrost.UserInfo{
-				ID:       "U-existing",
-				UserName: "ExistingUser",
-				IconUrl:  "https://example.test/icon.png",
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.CreateHeadlessAccountRequest{
-			Credential: "existing@example.test",
-			Password:   "newpassword123",
-		})
-
-		_, err := client.CreateHeadlessAccount(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_DeleteHeadlessAccount(t *testing.T) {
-	t.Run("成功: ヘッドレスアカウントを削除", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-todelete", "todelete@example.test", "password123")
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.DeleteHeadlessAccountRequest{
-			AccountId: "U-todelete",
-		})
-
-		res, err := client.DeleteHeadlessAccount(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-
-		// Verify account was deleted
-		_, err = setup.queries.GetHeadlessAccount(t.Context(), "U-todelete")
-		assert.Error(t, err)
-	})
-
-	t.Run("成功: 存在しないアカウントを削除（何も起こらない）", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// DeleteHeadlessAccountは:execで実装されているため、
-		// 存在しないアカウントを削除してもエラーにならない（何も削除されないだけ）
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.DeleteHeadlessAccountRequest{
-			AccountId: "U-nonexistent",
-		})
-
-		res, err := client.DeleteHeadlessAccount(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-}
-
-func TestControllerService_ListHeadlessHostImageTags(t *testing.T) {
-	t.Run("成功: イメージタグ一覧を取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Mock HostConnector to return test tags
-		setup.mockHostConnector.EXPECT().
-			ListContainerTags(gomock.Any(), nil).
-			Return(port.ContainerImageList{
-				{
-					Tag:             "2024.1.1-v1.0.0",
-					ResoniteVersion: "2024.1.1",
-					IsPreRelease:    false,
-					AppVersion:      "v1.0.0",
-				},
-				{
-					Tag:             "prerelease-2024.1.2-v1.1.0",
-					ResoniteVersion: "2024.1.2",
-					IsPreRelease:    true,
-					AppVersion:      "v1.1.0",
-				},
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListHeadlessHostImageTagsRequest{})
-
-		res, err := client.ListHeadlessHostImageTags(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Len(t, res.Msg.GetTags(), 2)
-
-		// Verify first tag
-		assert.Equal(t, "2024.1.1-v1.0.0", res.Msg.GetTags()[0].GetTag())
-		assert.Equal(t, "2024.1.1", res.Msg.GetTags()[0].GetResoniteVersion())
-		assert.False(t, res.Msg.GetTags()[0].GetIsPrerelease())
-		assert.Equal(t, "v1.0.0", res.Msg.GetTags()[0].GetAppVersion())
-
-		// Verify second tag
-		assert.Equal(t, "prerelease-2024.1.2-v1.1.0", res.Msg.GetTags()[1].GetTag())
-		assert.Equal(t, "2024.1.2", res.Msg.GetTags()[1].GetResoniteVersion())
-		assert.True(t, res.Msg.GetTags()[1].GetIsPrerelease())
-		assert.Equal(t, "v1.1.0", res.Msg.GetTags()[1].GetAppVersion())
-	})
-
-	t.Run("失敗: コネクタでエラー発生", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Mock HostConnector to return error
-		setup.mockHostConnector.EXPECT().
-			ListContainerTags(gomock.Any(), nil).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListHeadlessHostImageTagsRequest{})
-
-		_, err := client.ListHeadlessHostImageTags(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
 }
 
 func TestControllerService_Authentication(t *testing.T) {
@@ -477,3002 +201,5 @@ func TestControllerService_Authentication(t *testing.T) {
 		connectErr := new(connect.Error)
 		require.ErrorAs(t, err, &connectErr)
 		assert.Equal(t, connect.CodeUnauthenticated, connectErr.Code())
-	})
-}
-
-func TestControllerService_UpdateHeadlessAccountCredentials(t *testing.T) {
-	t.Run("成功: 認証情報を更新", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-update", "old@example.test", "oldpassword")
-
-		// Mock skyfrost client to return successful login with new credentials
-		setup.mockSkyfrost.EXPECT().
-			UserLogin(gomock.Any(), "new@example.test", "newpassword").
-			Return(&skyfrost.UserSession{UserId: "U-update"}, nil)
-
-		setup.mockSkyfrost.EXPECT().
-			FetchUserInfo(gomock.Any(), "U-update").
-			Return(&skyfrost.UserInfo{
-				ID:       "U-update",
-				UserName: "UpdatedUser",
-				IconUrl:  "https://example.test/icon.png",
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateHeadlessAccountCredentialsRequest{
-			AccountId:  "U-update",
-			Credential: "new@example.test",
-			Password:   "newpassword",
-		})
-
-		res, err := client.UpdateHeadlessAccountCredentials(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-
-		// Verify credentials were updated in DB
-		account, err := setup.queries.GetHeadlessAccount(t.Context(), "U-update")
-		require.NoError(t, err)
-		assert.Equal(t, "new@example.test", account.Credential)
-	})
-
-	t.Run("成功: 存在しないアカウントの認証情報を更新（何も起こらない）", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Mock skyfrost to return successful login
-		setup.mockSkyfrost.EXPECT().
-			UserLogin(gomock.Any(), "nonexist@example.test", "password").
-			Return(&skyfrost.UserSession{UserId: "U-nonexist"}, nil)
-
-		setup.mockSkyfrost.EXPECT().
-			FetchUserInfo(gomock.Any(), "U-nonexist").
-			Return(&skyfrost.UserInfo{
-				ID:       "U-nonexist",
-				UserName: "NonExist",
-				IconUrl:  "https://example.test/icon.png",
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateHeadlessAccountCredentialsRequest{
-			AccountId:  "U-nonexist",
-			Credential: "nonexist@example.test",
-			Password:   "password",
-		})
-
-		// UPDATE statement does not fail for non-existent records, it just updates 0 rows
-		_, err := client.UpdateHeadlessAccountCredentials(t.Context(), req)
-		require.NoError(t, err)
-	})
-
-	t.Run("失敗: 無効な新しい認証情報", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-invalidupdate", "valid@example.test", "validpassword")
-
-		// Mock skyfrost to return login error
-		setup.mockSkyfrost.EXPECT().
-			UserLogin(gomock.Any(), "invalid@example.test", "invalidpassword").
-			Return(nil, connect.NewError(connect.CodeUnauthenticated, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateHeadlessAccountCredentialsRequest{
-			AccountId:  "U-invalidupdate",
-			Credential: "invalid@example.test",
-			Password:   "invalidpassword",
-		})
-
-		_, err := client.UpdateHeadlessAccountCredentials(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_GetHeadlessAccountStorageInfo(t *testing.T) {
-	t.Run("成功: ストレージ情報を取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-storage-success", "user@example.test", "password")
-
-		// Mock skyfrost to return storage info
-		setup.mockSkyfrost.EXPECT().
-			GetStorageInfo(gomock.Any(), "user@example.test", "password", "U-storage-success").
-			Return(&skyfrost.StorageInfo{
-				UsedBytes:  1024 * 1024 * 100, // 100 MB
-				QuotaBytes: 1024 * 1024 * 500, // 500 MB
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessAccountStorageInfoRequest{
-			AccountId: "U-storage-success",
-		})
-
-		res, err := client.GetHeadlessAccountStorageInfo(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Equal(t, int64(1024*1024*100), res.Msg.GetStorageUsedBytes())
-		assert.Equal(t, int64(1024*1024*500), res.Msg.GetStorageQuotaBytes())
-	})
-
-	t.Run("失敗: 存在しないアカウント", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessAccountStorageInfoRequest{
-			AccountId: "U-nonexist",
-		})
-
-		_, err := client.GetHeadlessAccountStorageInfo(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-
-	t.Run("失敗: ストレージ情報の取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-storage", "user@example.test", "password")
-
-		// Mock skyfrost to return error
-		setup.mockSkyfrost.EXPECT().
-			GetStorageInfo(gomock.Any(), "user@example.test", "password", "U-storage").
-			Return(nil, connect.NewError(connect.CodeUnauthenticated, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessAccountStorageInfoRequest{
-			AccountId: "U-storage",
-		})
-
-		_, err := client.GetHeadlessAccountStorageInfo(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_RefetchHeadlessAccountInfo(t *testing.T) {
-	t.Run("成功: アカウント情報を再取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-refetch", "user@example.test", "password")
-
-		// Mock skyfrost to return user info
-		setup.mockSkyfrost.EXPECT().
-			FetchUserInfo(gomock.Any(), "U-refetch").
-			Return(&skyfrost.UserInfo{
-				ID:       "U-refetch",
-				UserName: "UpdatedName",
-				IconUrl:  "https://example.test/new-icon.png",
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.RefetchHeadlessAccountInfoRequest{
-			AccountId: "U-refetch",
-		})
-
-		res, err := client.RefetchHeadlessAccountInfo(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-
-		// Verify account info was updated
-		account, err := setup.queries.GetHeadlessAccount(t.Context(), "U-refetch")
-		require.NoError(t, err)
-		assert.Equal(t, "UpdatedName", account.LastDisplayName.String)
-		assert.Equal(t, "https://example.test/new-icon.png", account.LastIconUrl.String)
-	})
-
-	t.Run("失敗: 存在しないアカウント", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Mock skyfrost to return error for non-existent user
-		setup.mockSkyfrost.EXPECT().
-			FetchUserInfo(gomock.Any(), "U-nonexist").
-			Return(nil, connect.NewError(connect.CodeNotFound, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.RefetchHeadlessAccountInfoRequest{
-			AccountId: "U-nonexist",
-		})
-
-		_, err := client.RefetchHeadlessAccountInfo(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-
-	t.Run("失敗: ユーザー情報の取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-failfetch", "user@example.test", "password")
-
-		// Mock skyfrost to return error
-		setup.mockSkyfrost.EXPECT().
-			FetchUserInfo(gomock.Any(), "U-failfetch").
-			Return(nil, connect.NewError(connect.CodeNotFound, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.RefetchHeadlessAccountInfoRequest{
-			AccountId: "U-failfetch",
-		})
-
-		_, err := client.RefetchHeadlessAccountInfo(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_AcceptFriendRequests(t *testing.T) {
-	t.Run("成功: フレンドリクエストを承認", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-friend", "friend@example.test", "password")
-
-		// Create test host in database
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-friend", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector to return RPC client
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		// Mock RPC call to accept friend requests
-		setup.mockRpcClient.EXPECT().
-			AcceptFriendRequests(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.AcceptFriendRequestsResponse{}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.AcceptFriendRequestsRequest{
-			HeadlessAccountId: "U-friend",
-			TargetUserId:      "U-target",
-		})
-
-		res, err := client.AcceptFriendRequests(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-
-		// Verify host was actually created
-		_, err = setup.queries.GetHost(t.Context(), host.ID)
-		require.NoError(t, err)
-	})
-
-	t.Run("失敗: 起動中のホストがない", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-nohost", "nohost@example.test", "password")
-
-		// No running hosts created, so ListRunningByAccount will return empty list
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.AcceptFriendRequestsRequest{
-			HeadlessAccountId: "U-nohost",
-			TargetUserId:      "U-target",
-		})
-
-		_, err := client.AcceptFriendRequests(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInvalidArgument, connectErr.Code())
-	})
-
-	t.Run("失敗: RPCクライアントの取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-rpcfail", "rpcfail@example.test", "password")
-
-		// Create test host
-		testutil.CreateTestHeadlessHost(t, setup.queries, "U-rpcfail", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector to return error when getting RPC client
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.AcceptFriendRequestsRequest{
-			HeadlessAccountId: "U-rpcfail",
-			TargetUserId:      "U-target",
-		})
-
-		_, err := client.AcceptFriendRequests(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_GetFriendRequests(t *testing.T) {
-	t.Run("成功: フレンドリクエストを取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-friend", "friend@example.test", "password")
-
-		// Mock skyfrost to return friend requests
-		setup.mockSkyfrost.EXPECT().
-			GetContacts(gomock.Any(), "friend@example.test", "password").
-			Return([]skyfrost.Contact{
-				{
-					Id:         "contact1",
-					Username:   "User1",
-					Status:     "Requested",
-					IsAccepted: false,
-				},
-				{
-					Id:         "contact2",
-					Username:   "User2",
-					Status:     "Requested",
-					IsAccepted: false,
-				},
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetFriendRequestsRequest{
-			HeadlessAccountId: "U-friend",
-		})
-
-		res, err := client.GetFriendRequests(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Len(t, res.Msg.GetRequestedContacts(), 2)
-	})
-
-	t.Run("失敗: 存在しないアカウント", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetFriendRequestsRequest{
-			HeadlessAccountId: "U-nonexist",
-		})
-
-		_, err := client.GetFriendRequests(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-
-	t.Run("失敗: コンタクト取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-contactfail", "user@example.test", "password")
-
-		// Mock skyfrost to return contacts error
-		setup.mockSkyfrost.EXPECT().
-			GetContacts(gomock.Any(), "user@example.test", "password").
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetFriendRequestsRequest{
-			HeadlessAccountId: "U-contactfail",
-		})
-
-		_, err := client.GetFriendRequests(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_StartHeadlessHost(t *testing.T) {
-	t.Run("成功: ホストを起動", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-
-		setup.mockHostConnector.EXPECT().
-			Start(gomock.Any(), gomock.Any()).
-			DoAndReturn(func(ctx context.Context, params hostconnector.HostStartParams) (hostconnector.HostConnectString, error) {
-				assert.Equal(t, int32(1), params.InstanceId, "Initial start should have InstanceId=1")
-
-				return hostconnector.HostConnectString("test-container"), nil
-			})
-
-		imageTag := "latest"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.StartHeadlessHostRequest{
-			HeadlessAccountId: "U-test",
-			Name:              "TestHost",
-			ImageTag:          &imageTag,
-		})
-
-		res, err := client.StartHeadlessHost(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.NotEmpty(t, res.Msg.GetHostId())
-
-		// Verify host was created in database
-		host, err := setup.queries.GetHost(t.Context(), res.Msg.GetHostId())
-		require.NoError(t, err)
-		assert.Equal(t, "TestHost", host.Name)
-		assert.Equal(t, "U-test", host.AccountID)
-		assert.Equal(t, int32(1), host.InstanceCount, "Instance count should be 1 after first start")
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil).
-			AnyTimes()
-
-		setup.mockRpcClient.EXPECT().
-			GetAbout(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetAboutResponse{
-				ResoniteVersion: "1.0.0",
-				AppVersion:      "1.0.0",
-			}, nil).
-			AnyTimes()
-
-		setup.mockRpcClient.EXPECT().
-			GetAccountInfo(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetAccountInfoResponse{
-				UserId:      "U-test",
-				DisplayName: "Test Account",
-			}, nil).
-			AnyTimes()
-
-		setup.mockRpcClient.EXPECT().
-			GetStatus(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetStatusResponse{
-				Fps: 60.0,
-			}, nil).
-			AnyTimes()
-
-		setup.mockRpcClient.EXPECT().
-			GetStartupConfigToRestore(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetStartupConfigToRestoreResponse{
-				StartupConfig: &headlessv1.StartupConfig{},
-			}, nil).
-			AnyTimes()
-
-		getReq := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostRequest{
-			HostId: res.Msg.GetHostId(),
-		})
-
-		getRes, err := client.GetHeadlessHost(t.Context(), getReq)
-		require.NoError(t, err)
-		assert.Equal(t, int32(1), getRes.Msg.GetHost().GetInstanceId(), "RPC response should include instance_id=1")
-	})
-
-	t.Run("失敗: 存在しないアカウント", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		imageTag := "latest"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.StartHeadlessHostRequest{
-			HeadlessAccountId: "U-nonexist",
-			Name:              "TestHost",
-			ImageTag:          &imageTag,
-		})
-
-		_, err := client.StartHeadlessHost(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_RestartHeadlessHost(t *testing.T) {
-	t.Run("成功: ホストを再起動", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector - ListContainerTags (called by resolveTagToUse)
-		setup.mockHostConnector.EXPECT().
-			ListContainerTags(gomock.Any(), gomock.Any()).
-			Return(port.ContainerImageList{
-				{Tag: "latest", IsPreRelease: false},
-			}, nil).
-			Times(1)
-
-		// Mock HostConnector - GetRpcClient (may be called multiple times)
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil).
-			AnyTimes()
-
-		// Mock RPC client to return about info
-		setup.mockRpcClient.EXPECT().
-			GetAbout(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetAboutResponse{
-				ResoniteVersion: "1.0.0",
-				AppVersion:      "1.0.0",
-			}, nil).
-			AnyTimes()
-
-		// Mock RPC client to return account info
-		setup.mockRpcClient.EXPECT().
-			GetAccountInfo(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetAccountInfoResponse{
-				UserId:      "U-test",
-				DisplayName: "Test Account",
-			}, nil).
-			AnyTimes()
-
-		// Mock RPC client to return status
-		setup.mockRpcClient.EXPECT().
-			GetStatus(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetStatusResponse{
-				Fps: 60.0,
-			}, nil).
-			AnyTimes()
-
-		// Mock RPC client to return startup config
-		setup.mockRpcClient.EXPECT().
-			GetStartupConfigToRestore(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetStartupConfigToRestoreResponse{
-				StartupConfig: &headlessv1.StartupConfig{},
-			}, nil).
-			AnyTimes()
-
-		// Mock RPC client to return sessions (for SearchSessions)
-		setup.mockRpcClient.EXPECT().
-			ListSessions(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.ListSessionsResponse{
-				Sessions: []*headlessv1.Session{},
-			}, nil).
-			AnyTimes()
-
-		// Mock HostConnector - Stop
-		setup.mockHostConnector.EXPECT().
-			Stop(gomock.Any(), gomock.Any(), gomock.Any()).
-			Return(nil).
-			Times(1)
-
-		// 旧コンテナの削除 (issue #21 対応)
-		setup.mockHostConnector.EXPECT().
-			Remove(gomock.Any(), gomock.Any()).
-			Return(nil).
-			Times(1)
-
-		setup.mockHostConnector.EXPECT().
-			Start(gomock.Any(), gomock.Any()).
-			DoAndReturn(func(ctx context.Context, params hostconnector.HostStartParams) (hostconnector.HostConnectString, error) {
-				assert.Equal(t, int32(2), params.InstanceId, "Restart should have InstanceId=2")
-
-				return hostconnector.HostConnectString("test-container-restarted"), nil
-			}).
-			Times(1)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.RestartHeadlessHostRequest{
-			HostId: host.ID,
-		})
-
-		res, err := client.RestartHeadlessHost(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-
-		updatedHost, err := setup.queries.GetHost(t.Context(), host.ID)
-		require.NoError(t, err)
-		assert.Equal(t, int32(2), updatedHost.InstanceCount, "Instance count should be 2 after restart")
-	})
-
-	t.Run("失敗: 存在しないホスト", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.RestartHeadlessHostRequest{
-			HostId: "nonexist-host",
-		})
-
-		_, err := client.RestartHeadlessHost(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeNotFound, connectErr.Code())
-	})
-}
-
-func TestControllerService_UpdateHeadlessHostSettings(t *testing.T) {
-	t.Run("成功: 停止中のホストの設定を更新", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_EXITED)
-
-		newName := "UpdatedHost"
-		newTickRate := float32(120)
-		newPolicy := hdlctrlv1.HeadlessHostAutoUpdatePolicy_HEADLESS_HOST_AUTO_UPDATE_POLICY_USERS_EMPTY
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateHeadlessHostSettingsRequest{
-			HostId:           host.ID,
-			Name:             &newName,
-			TickRate:         &newTickRate,
-			AutoUpdatePolicy: &newPolicy,
-		})
-
-		res, err := client.UpdateHeadlessHostSettings(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-
-		// Verify host was updated in database
-		updatedHost, err := setup.queries.GetHost(t.Context(), host.ID)
-		require.NoError(t, err)
-		assert.Equal(t, newName, updatedHost.Name)
-		assert.Equal(t, int32(entity.HostAutoUpdatePolicy_USERS_EMPTY), updatedHost.AutoUpdatePolicy,
-			"AutoUpdatePolicy passed in request should be persisted")
-	})
-
-	t.Run("成功: 実行中のホストの設定を更新", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and running host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test2", "test2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test2", "RunningHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock RPC calls - GetRpcClient is called twice: once in dbToEntity, once before UpdateHostSettings
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil).
-			Times(2)
-
-		newTickRate := float32(120)
-		maxTransfers := int32(4)
-
-		setup.mockRpcClient.EXPECT().
-			GetAccountInfo(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetAccountInfoResponse{}, nil)
-
-		setup.mockRpcClient.EXPECT().
-			GetStatus(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetStatusResponse{}, nil)
-
-		setup.mockRpcClient.EXPECT().
-			GetAbout(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetAboutResponse{}, nil)
-
-		// GetStartupConfigToRestore is called twice: once in dbToEntity, once after UpdateHostSettings
-		setup.mockRpcClient.EXPECT().
-			GetStartupConfigToRestore(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetStartupConfigToRestoreResponse{
-				StartupConfig: &headlessv1.StartupConfig{
-					TickRate:                    &newTickRate,
-					MaxConcurrentAssetTransfers: &maxTransfers,
-				},
-			}, nil).
-			Times(2)
-
-		setup.mockRpcClient.EXPECT().
-			UpdateHostSettings(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.UpdateHostSettingsResponse{}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateHeadlessHostSettingsRequest{
-			HostId:   host.ID,
-			TickRate: &newTickRate,
-		})
-
-		res, err := client.UpdateHeadlessHostSettings(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("失敗: 存在しないホスト", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateHeadlessHostSettingsRequest{
-			HostId: "nonexist",
-		})
-
-		_, err := client.UpdateHeadlessHostSettings(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeNotFound, connectErr.Code())
-	})
-}
-
-func TestControllerService_GetHeadlessHostLogs(t *testing.T) {
-	t.Run("成功: ログを取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_EXITED)
-
-		// Insert test logs into container_logs table
-		baseTime := time.Date(2024, 1, 15, 10, 30, 0, 0, time.UTC)
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, host.InstanceCount, baseTime, "stdout", "Log line 1")
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, host.InstanceCount, baseTime.Add(time.Second), "stderr", "Error line")
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostLogsRequest{
-			HostId:     host.ID,
-			InstanceId: host.InstanceCount,
-		})
-
-		res, err := client.GetHeadlessHostLogs(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Len(t, res.Msg.GetLogs(), 2)
-		assert.Equal(t, "Log line 1", res.Msg.GetLogs()[0].GetBody())
-		assert.False(t, res.Msg.GetLogs()[0].GetIsError())
-		assert.Equal(t, "Error line", res.Msg.GetLogs()[1].GetBody())
-		assert.True(t, res.Msg.GetLogs()[1].GetIsError())
-	})
-
-	t.Run("成功: ログが存在しない場合は空のリストを返す", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host (no logs inserted)
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test2", "test2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test2", "TestHost2", entity.HeadlessHostStatus_EXITED)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostLogsRequest{
-			HostId:     host.ID,
-			InstanceId: host.InstanceCount,
-		})
-
-		res, err := client.GetHeadlessHostLogs(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Empty(t, res.Msg.GetLogs())
-	})
-
-	t.Run("成功: limitパラメータでログ件数を制限", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test3", "test3@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test3", "TestHost3", entity.HeadlessHostStatus_EXITED)
-
-		// Insert 5 logs
-		baseTime := time.Date(2024, 1, 15, 10, 30, 0, 0, time.UTC)
-		for i := range 5 {
-			testutil.InsertTestContainerLog(t, setup.queries, host.ID, host.InstanceCount, baseTime.Add(time.Duration(i)*time.Second), "stdout", fmt.Sprintf("Log line %d", i+1))
-		}
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostLogsRequest{
-			HostId:     host.ID,
-			InstanceId: host.InstanceCount,
-			Limit:      3,
-		})
-
-		res, err := client.GetHeadlessHostLogs(t.Context(), req)
-		require.NoError(t, err)
-		assert.Len(t, res.Msg.GetLogs(), 3)
-		assert.True(t, res.Msg.GetHasMoreBefore(), "should have more logs before")
-		assert.False(t, res.Msg.GetHasMoreAfter(), "should not have more logs after (initial fetch)")
-	})
-
-	t.Run("成功: beforeIdカーソルで古いログを取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test4", "test4@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test4", "TestHost4", entity.HeadlessHostStatus_EXITED)
-
-		// Insert logs with distinct timestamps
-		baseTime := time.Date(2024, 1, 15, 10, 30, 0, 0, time.UTC)
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, host.InstanceCount, baseTime, "stdout", "Old log")
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, host.InstanceCount, baseTime.Add(time.Minute), "stdout", "Middle log")
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, host.InstanceCount, baseTime.Add(2*time.Minute), "stdout", "New log")
-
-		// First, get all logs to find the ID of "Middle log"
-		allLogsReq := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostLogsRequest{
-			HostId:     host.ID,
-			InstanceId: host.InstanceCount,
-		})
-		allLogsRes, err := client.GetHeadlessHostLogs(t.Context(), allLogsReq)
-		require.NoError(t, err)
-		require.Len(t, allLogsRes.Msg.GetLogs(), 3)
-
-		// Find the middle log's ID (logs are returned in chronological order: Old, Middle, New)
-		var middleLogId int64
-
-		for _, log := range allLogsRes.Msg.GetLogs() {
-			if log.GetBody() == "Middle log" {
-				middleLogId = log.GetId()
-
-				break
-			}
-		}
-
-		require.NotZero(t, middleLogId, "should find middle log")
-
-		// Use beforeId cursor to get logs before "Middle log"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostLogsRequest{
-			HostId:     host.ID,
-			InstanceId: host.InstanceCount,
-			Cursor:     &hdlctrlv1.GetHeadlessHostLogsRequest_BeforeId{BeforeId: middleLogId},
-		})
-
-		res, err := client.GetHeadlessHostLogs(t.Context(), req)
-		require.NoError(t, err)
-		assert.Len(t, res.Msg.GetLogs(), 1)
-		assert.Equal(t, "Old log", res.Msg.GetLogs()[0].GetBody())
-		assert.False(t, res.Msg.GetHasMoreBefore(), "no more older logs")
-	})
-
-	t.Run("成功: afterIdカーソルで新しいログを取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test5", "test5@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test5", "TestHost5", entity.HeadlessHostStatus_EXITED)
-
-		// Insert logs with distinct timestamps
-		baseTime := time.Date(2024, 1, 15, 10, 30, 0, 0, time.UTC)
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, host.InstanceCount, baseTime, "stdout", "Old log")
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, host.InstanceCount, baseTime.Add(time.Minute), "stdout", "Middle log")
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, host.InstanceCount, baseTime.Add(2*time.Minute), "stdout", "New log")
-
-		// First, get all logs to find the ID of "Middle log"
-		allLogsReq := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostLogsRequest{
-			HostId:     host.ID,
-			InstanceId: host.InstanceCount,
-		})
-		allLogsRes, err := client.GetHeadlessHostLogs(t.Context(), allLogsReq)
-		require.NoError(t, err)
-		require.Len(t, allLogsRes.Msg.GetLogs(), 3)
-
-		// Find the middle log's ID (logs are returned in chronological order: Old, Middle, New)
-		var middleLogId int64
-
-		for _, log := range allLogsRes.Msg.GetLogs() {
-			if log.GetBody() == "Middle log" {
-				middleLogId = log.GetId()
-
-				break
-			}
-		}
-
-		require.NotZero(t, middleLogId, "should find middle log")
-
-		// Use afterId cursor to get logs after "Middle log"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostLogsRequest{
-			HostId:     host.ID,
-			InstanceId: host.InstanceCount,
-			Cursor:     &hdlctrlv1.GetHeadlessHostLogsRequest_AfterId{AfterId: middleLogId},
-		})
-
-		res, err := client.GetHeadlessHostLogs(t.Context(), req)
-		require.NoError(t, err)
-		assert.Len(t, res.Msg.GetLogs(), 1)
-		assert.Equal(t, "New log", res.Msg.GetLogs()[0].GetBody())
-		assert.False(t, res.Msg.GetHasMoreAfter(), "no more newer logs")
-	})
-
-	t.Run("成功: 異なるinstanceIdでログを分離", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test6", "test6@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test6", "TestHost6", entity.HeadlessHostStatus_EXITED)
-
-		baseTime := time.Date(2024, 1, 15, 10, 30, 0, 0, time.UTC)
-		// Insert logs for instance 1 (current)
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, host.InstanceCount, baseTime, "stdout", "Current instance log")
-		// Insert logs for instance 0 (previous)
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, 0, baseTime, "stdout", "Previous instance log")
-
-		// Request current instance logs
-		req1 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostLogsRequest{
-			HostId:     host.ID,
-			InstanceId: host.InstanceCount,
-		})
-		res1, err := client.GetHeadlessHostLogs(t.Context(), req1)
-		require.NoError(t, err)
-		assert.Len(t, res1.Msg.GetLogs(), 1)
-		assert.Equal(t, "Current instance log", res1.Msg.GetLogs()[0].GetBody())
-
-		// Request previous instance logs
-		req2 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostLogsRequest{
-			HostId:     host.ID,
-			InstanceId: 0,
-		})
-		res2, err := client.GetHeadlessHostLogs(t.Context(), req2)
-		require.NoError(t, err)
-		assert.Len(t, res2.Msg.GetLogs(), 1)
-		assert.Equal(t, "Previous instance log", res2.Msg.GetLogs()[0].GetBody())
-	})
-
-	t.Run("成功: has_more_afterフラグが正しく設定される", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test7", "test7@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test7", "TestHost7", entity.HeadlessHostStatus_EXITED)
-
-		// Insert 5 logs
-		baseTime := time.Date(2024, 1, 15, 10, 30, 0, 0, time.UTC)
-		for i := range 5 {
-			testutil.InsertTestContainerLog(t, setup.queries, host.ID, host.InstanceCount, baseTime.Add(time.Duration(i)*time.Minute), "stdout", fmt.Sprintf("Log %d", i+1))
-		}
-
-		// First, get all logs to find the ID of "Log 1" (oldest)
-		allLogsReq := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostLogsRequest{
-			HostId:     host.ID,
-			InstanceId: host.InstanceCount,
-		})
-		allLogsRes, err := client.GetHeadlessHostLogs(t.Context(), allLogsReq)
-		require.NoError(t, err)
-		require.Len(t, allLogsRes.Msg.GetLogs(), 5)
-
-		// Get the first log's ID (oldest, "Log 1")
-		firstLogId := allLogsRes.Msg.GetLogs()[0].GetId()
-		require.NotZero(t, firstLogId, "should have first log ID")
-
-		// Use afterId cursor with limit to trigger has_more_after
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostLogsRequest{
-			HostId:     host.ID,
-			InstanceId: host.InstanceCount,
-			Limit:      2,
-			Cursor:     &hdlctrlv1.GetHeadlessHostLogsRequest_AfterId{AfterId: firstLogId},
-		})
-
-		res, err := client.GetHeadlessHostLogs(t.Context(), req)
-		require.NoError(t, err)
-		assert.Len(t, res.Msg.GetLogs(), 2)
-		assert.True(t, res.Msg.GetHasMoreAfter(), "should have more logs after")
-		assert.False(t, res.Msg.GetHasMoreBefore(), "should not have has_more_before with after cursor")
-	})
-}
-
-func TestControllerService_ShutdownHeadlessHost(t *testing.T) {
-	t.Run("成功: ホストをシャットダウン", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector - Stop calls GetRpcClient to fetch startup config
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			GetStartupConfigToRestore(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetStartupConfigToRestoreResponse{
-				StartupConfig: &headlessv1.StartupConfig{},
-			}, nil)
-
-		setup.mockHostConnector.EXPECT().
-			Stop(gomock.Any(), gomock.Any(), gomock.Any()).
-			Return(nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ShutdownHeadlessHostRequest{
-			HostId: host.ID,
-		})
-
-		res, err := client.ShutdownHeadlessHost(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("失敗: 存在しないホスト", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ShutdownHeadlessHostRequest{
-			HostId: "nonexist",
-		})
-
-		_, err := client.ShutdownHeadlessHost(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeNotFound, connectErr.Code())
-	})
-}
-
-func TestControllerService_KillHeadlessHost(t *testing.T) {
-	t.Run("成功: ホストを強制停止", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector - Kill
-		setup.mockHostConnector.EXPECT().
-			Kill(gomock.Any(), gomock.Any()).
-			Return(nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.KillHeadlessHostRequest{
-			HostId: host.ID,
-		})
-
-		res, err := client.KillHeadlessHost(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("失敗: 存在しないホスト", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.KillHeadlessHostRequest{
-			HostId: "nonexist",
-		})
-
-		_, err := client.KillHeadlessHost(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeNotFound, connectErr.Code())
-	})
-}
-
-func TestControllerService_GetHeadlessHost(t *testing.T) {
-	t.Run("成功: ホストの詳細を取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_EXITED)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostRequest{
-			HostId: host.ID,
-		})
-
-		res, err := client.GetHeadlessHost(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.NotNil(t, res.Msg.GetHost())
-		assert.Equal(t, host.ID, res.Msg.GetHost().GetId())
-		assert.Equal(t, "TestHost", res.Msg.GetHost().GetName())
-	})
-
-	t.Run("失敗: 存在しないホスト", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetHeadlessHostRequest{
-			HostId: "nonexist",
-		})
-
-		_, err := client.GetHeadlessHost(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeNotFound, connectErr.Code())
-	})
-}
-
-func TestControllerService_ListHeadlessHost(t *testing.T) {
-	t.Run("成功: ホスト一覧を取得 (ページング検証)", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create 5 test hosts for pagination
-		const totalHosts = 5
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-
-		for i := 1; i <= totalHosts; i++ {
-			testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", fmt.Sprintf("Host%d", i), entity.HeadlessHostStatus_EXITED)
-		}
-
-		// page 未指定 -> デフォルト 20 件 (5件しかないので全件返る)
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListHeadlessHostRequest{})
-		res, err := client.ListHeadlessHost(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Len(t, res.Msg.GetHosts(), totalHosts)
-		require.NotNil(t, res.Msg.GetPage())
-		assert.Equal(t, int32(totalHosts), res.Msg.GetPage().GetTotalCount())
-		assert.Equal(t, int32(20), res.Msg.GetPage().GetPageSize())
-
-		// page_size=3 で 1 ページ目 -> 3 件
-		req2 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListHeadlessHostRequest{
-			Page: &hdlctrlv1.PageRequest{PageIndex: 0, PageSize: 3},
-		})
-		res2, err := client.ListHeadlessHost(t.Context(), req2)
-		require.NoError(t, err)
-		assert.Len(t, res2.Msg.GetHosts(), 3)
-		assert.Equal(t, int32(totalHosts), res2.Msg.GetPage().GetTotalCount())
-
-		// page_size=3, page_index=1 -> 残り 2 件
-		req3 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListHeadlessHostRequest{
-			Page: &hdlctrlv1.PageRequest{PageIndex: 1, PageSize: 3},
-		})
-		res3, err := client.ListHeadlessHost(t.Context(), req3)
-		require.NoError(t, err)
-		assert.Len(t, res3.Msg.GetHosts(), 2)
-		assert.Equal(t, int32(totalHosts), res3.Msg.GetPage().GetTotalCount())
-
-		// page_index=-1 -> CodeInvalidArgument
-		req4 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListHeadlessHostRequest{
-			Page: &hdlctrlv1.PageRequest{PageIndex: -1, PageSize: 20},
-		})
-		_, err = client.ListHeadlessHost(t.Context(), req4)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok)
-		assert.Equal(t, connect.CodeInvalidArgument, connectErr.Code())
-	})
-}
-
-func TestControllerService_DeleteHeadlessHost(t *testing.T) {
-	t.Run("成功: ホストを削除", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_EXITED)
-
-		// Add container logs for this host (multiple instances)
-		now := time.Now()
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, 1, now, "stdout", "log message 1")
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, 1, now.Add(time.Second), "stderr", "error message")
-		testutil.InsertTestContainerLog(t, setup.queries, host.ID, 2, now, "stdout", "log from instance 2")
-
-		// Verify logs exist before deletion
-		logsBefore, err := setup.queries.GetContainerLogsByTag(t.Context(), db.GetContainerLogsByTagParams{
-			Tag:     pgtype.Text{String: "headless-" + host.ID + "-1", Valid: true},
-			MaxRows: int32(100),
-		})
-		require.NoError(t, err)
-		assert.Len(t, logsBefore, 2, "should have 2 logs for instance 1 before deletion")
-
-		// Mock the container removal
-		setup.mockHostConnector.EXPECT().
-			Remove(gomock.Any(), gomock.Any()).
-			Return(nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.DeleteHeadlessHostRequest{
-			HostId: host.ID,
-		})
-
-		res, err := client.DeleteHeadlessHost(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-
-		// Verify host was deleted from database
-		_, err = setup.queries.GetHost(t.Context(), host.ID)
-		require.Error(t, err)
-
-		// Verify container logs were also deleted
-		logsAfter1, err := setup.queries.GetContainerLogsByTag(t.Context(), db.GetContainerLogsByTagParams{
-			Tag:     pgtype.Text{String: "headless-" + host.ID + "-1", Valid: true},
-			MaxRows: int32(100),
-		})
-		require.NoError(t, err)
-		assert.Empty(t, logsAfter1, "logs for instance 1 should be deleted")
-
-		logsAfter2, err := setup.queries.GetContainerLogsByTag(t.Context(), db.GetContainerLogsByTagParams{
-			Tag:     pgtype.Text{String: "headless-" + host.ID + "-2", Valid: true},
-			MaxRows: int32(100),
-		})
-		require.NoError(t, err)
-		assert.Empty(t, logsAfter2, "logs for instance 2 should be deleted")
-	})
-
-	t.Run("成功: 存在しないホストを削除（何も起こらない）", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.DeleteHeadlessHostRequest{
-			HostId: "nonexist",
-		})
-
-		res, err := client.DeleteHeadlessHost(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-}
-
-func TestControllerService_AllowHostAccess(t *testing.T) {
-	t.Run("成功: ホストアクセスを許可", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			AllowHostAccess(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.AllowHostAccessResponse{}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.AllowHostAccessRequest{
-			HostId:  host.ID,
-			Request: &headlessv1.AllowHostAccessRequest{},
-		})
-
-		res, err := client.AllowHostAccess(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("失敗: RPCクライアントの取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test2", "test2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test2", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.AllowHostAccessRequest{
-			HostId: host.ID,
-
-			Request: &headlessv1.AllowHostAccessRequest{},
-		})
-
-		_, err := client.AllowHostAccess(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-
-	t.Run("失敗: RPC呼び出しに失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test3", "test3@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test3", "TestHost3", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			AllowHostAccess(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.AllowHostAccessRequest{
-			HostId:  host.ID,
-			Request: &headlessv1.AllowHostAccessRequest{},
-		})
-
-		_, err := client.AllowHostAccess(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_DenyHostAccess(t *testing.T) {
-	t.Run("成功: ホストアクセスを拒否", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			DenyHostAccess(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.DenyHostAccessResponse{}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.DenyHostAccessRequest{
-			HostId:  host.ID,
-			Request: &headlessv1.DenyHostAccessRequest{},
-		})
-
-		res, err := client.DenyHostAccess(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("失敗: RPCクライアントの取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test2", "test2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test2", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.DenyHostAccessRequest{
-			HostId:  host.ID,
-			Request: &headlessv1.DenyHostAccessRequest{},
-		})
-
-		_, err := client.DenyHostAccess(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_BanUser(t *testing.T) {
-	t.Run("成功: ユーザーをBANする", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			BanUser(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.BanUserResponse{}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.BanUserRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.BanUserRequest{
-				User: &headlessv1.BanUserRequest_UserId{UserId: "U-target"},
-			},
-		})
-
-		res, err := client.BanUser(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("失敗: RPCクライアントの取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test2", "test2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test2", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.BanUserRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.BanUserRequest{
-				User: &headlessv1.BanUserRequest_UserId{UserId: "U-target"},
-			},
-		})
-
-		_, err := client.BanUser(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_KickUser(t *testing.T) {
-	t.Run("成功: ユーザーをキックする", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			KickUser(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.KickUserResponse{}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.KickUserRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.KickUserRequest{
-				User: &headlessv1.KickUserRequest_UserId{UserId: "U-target"},
-			},
-		})
-
-		res, err := client.KickUser(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("失敗: RPCクライアントの取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test2", "test2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test2", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.KickUserRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.KickUserRequest{
-				User: &headlessv1.KickUserRequest_UserId{UserId: "U-target"},
-			},
-		})
-
-		_, err := client.KickUser(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_SearchUserInfo(t *testing.T) {
-	t.Run("成功: ユーザー情報を検索", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			SearchUserInfo(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.SearchUserInfoResponse{
-				Users: []*headlessv1.UserInfo{
-					{
-						Id:   "U-found",
-						Name: "FoundUser",
-					},
-				},
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SearchUserInfoRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.SearchUserInfoRequest{
-				User: &headlessv1.SearchUserInfoRequest_UserName{UserName: "test"},
-			},
-		})
-
-		res, err := client.SearchUserInfo(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Len(t, res.Msg.GetUsers(), 1)
-		assert.Equal(t, "U-found", res.Msg.GetUsers()[0].GetId())
-	})
-
-	t.Run("失敗: RPCクライアントの取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test2", "test2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test2", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SearchUserInfoRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.SearchUserInfoRequest{
-				User: &headlessv1.SearchUserInfoRequest_UserName{UserName: "test"},
-			},
-		})
-
-		_, err := client.SearchUserInfo(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_FetchWorldInfo(t *testing.T) {
-	t.Run("成功: ワールド情報を取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			FetchWorldInfo(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.FetchWorldInfoResponse{
-				Name: "TestWorld",
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.FetchWorldInfoRequest{
-			HostId: host.ID,
-			Url:    "resrec:///U-test/R-12345",
-		})
-
-		res, err := client.FetchWorldInfo(t.Context(), req)
-
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Equal(t, "TestWorld", res.Msg.GetName())
-	})
-
-	t.Run("失敗: RPCクライアントの取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test2", "test2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test2", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.FetchWorldInfoRequest{
-			HostId: host.ID,
-			Url:    "resrec:///U-test/R-12345",
-		})
-
-		_, err := client.FetchWorldInfo(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-
-	t.Run("失敗: ワールドが見つからない", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test3", "test3@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test3", "TestHost3", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			FetchWorldInfo(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeNotFound, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.FetchWorldInfoRequest{
-			HostId: host.ID,
-			Url:    "resrec:///U-test/R-nonexist",
-		})
-
-		_, err := client.FetchWorldInfo(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeNotFound, connectErr.Code())
-	})
-}
-
-func TestControllerService_GetSessionDetails(t *testing.T) {
-	t.Run("成功: セッション詳細を取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Create test session
-		session := testutil.CreateTestSession(t, setup.queries, host.ID, "TestSession", entity.SessionStatus_RUNNING)
-
-		// Mock HostConnector - GetRpcClient
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil).
-			AnyTimes()
-
-		// Mock RPC call to get session
-		setup.mockRpcClient.EXPECT().
-			GetSession(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetSessionResponse{
-				Session: &headlessv1.Session{
-					Id:   session.ID,
-					Name: session.Name,
-				},
-			}, nil).
-			AnyTimes()
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetSessionDetailsRequest{
-			SessionId: session.ID,
-		})
-
-		res, err := client.GetSessionDetails(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Equal(t, session.ID, res.Msg.GetSession().GetId())
-		assert.Equal(t, "TestSession", res.Msg.GetSession().GetName())
-	})
-
-	t.Run("失敗: 存在しないセッション", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetSessionDetailsRequest{
-			SessionId: "nonexist-session",
-		})
-
-		_, err := client.GetSessionDetails(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeNotFound, connectErr.Code())
-	})
-}
-
-func TestControllerService_ListUsersInSession(t *testing.T) {
-	t.Run("成功: セッション内のユーザーリストを取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			ListUsersInSession(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.ListUsersInSessionResponse{
-				Users: []*headlessv1.UserInSession{
-					{Id: "U-user1", Name: "User1"},
-					{Id: "U-user2", Name: "User2"},
-				},
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListUsersInSessionRequest{
-			HostId:    host.ID,
-			SessionId: "session-123",
-		})
-
-		res, err := client.ListUsersInSession(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Len(t, res.Msg.GetUsers(), 2)
-		assert.Equal(t, "U-user1", res.Msg.GetUsers()[0].GetId())
-	})
-
-	t.Run("失敗: RPCクライアントの取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test2", "test2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test2", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListUsersInSessionRequest{
-			HostId:    host.ID,
-			SessionId: "session-123",
-		})
-
-		_, err := client.ListUsersInSession(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_SaveSessionWorld(t *testing.T) {
-	t.Run("成功: セッションのワールドを保存", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account, host, and session
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-		session := testutil.CreateTestSession(t, setup.queries, host.ID, "TestSession", entity.SessionStatus_RUNNING)
-
-		// Mock HostConnector - GetRpcClient
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil).
-			AnyTimes()
-
-		// SessionUsecase.GetSession は cache miss 時に container の GetSession を打って
-		// cache を populate する。テストでは cache 未投入なのでこの呼び出しが発生する。
-		setup.mockRpcClient.EXPECT().
-			GetSession(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetSessionResponse{
-				Session: &headlessv1.Session{Id: session.ID, Name: session.Name},
-			}, nil).
-			AnyTimes()
-
-		// Mock RPC call to save world. container 側が新しく saved_world_url を返すので
-		// preset 由来の初回 save でも stale な CurrentState ではなく同期的に正しい URL を返す。
-		setup.mockRpcClient.EXPECT().
-			SaveSessionWorld(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.SaveSessionWorldResponse{
-				SavedWorldUrl: "resrec:///U-test/R-test-world",
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SaveSessionWorldRequest{
-			SessionId: session.ID,
-			SaveMode:  hdlctrlv1.SaveSessionWorldRequest_SAVE_MODE_OVERWRITE,
-		})
-
-		res, err := client.SaveSessionWorld(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.NotNil(t, res.Msg.GetSavedRecordUrl())
-		assert.Equal(t, "resrec:///U-test/R-test-world", res.Msg.GetSavedRecordUrl())
-	})
-
-	t.Run("失敗: 存在しないセッション", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SaveSessionWorldRequest{
-			SessionId: "nonexist-session",
-			SaveMode:  hdlctrlv1.SaveSessionWorldRequest_SAVE_MODE_OVERWRITE,
-		})
-
-		_, err := client.SaveSessionWorld(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-
-	t.Run("失敗: 無効なセーブモード", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SaveSessionWorldRequest{
-			SessionId: "session-123",
-			SaveMode:  hdlctrlv1.SaveSessionWorldRequest_SAVE_MODE_UNKNOWN,
-		})
-
-		_, err := client.SaveSessionWorld(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInvalidArgument, connectErr.Code())
-	})
-}
-
-func TestControllerService_UpdateSessionParameters(t *testing.T) {
-	t.Run("成功: セッションパラメータを更新", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account, host, and session
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-		session := testutil.CreateTestSession(t, setup.queries, host.ID, "TestSession", entity.SessionStatus_RUNNING)
-
-		// Mock HostConnector - GetRpcClient
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil).
-			AnyTimes()
-
-		// Mock RPC call to get session (called by usecase)
-		setup.mockRpcClient.EXPECT().
-			GetSession(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetSessionResponse{
-				Session: &headlessv1.Session{
-					Id:   session.ID,
-					Name: session.Name,
-				},
-			}, nil).
-			AnyTimes()
-
-		// Mock RPC call to update parameters
-		maxUsers := int32(16)
-
-		setup.mockRpcClient.EXPECT().
-			UpdateSessionParameters(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.UpdateSessionParametersResponse{}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateSessionParametersRequest{
-			Parameters: &headlessv1.UpdateSessionParametersRequest{
-				SessionId: session.ID,
-				MaxUsers:  &maxUsers,
-			},
-		})
-
-		res, err := client.UpdateSessionParameters(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("失敗: 存在しないセッション", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateSessionParametersRequest{
-			Parameters: &headlessv1.UpdateSessionParametersRequest{
-				SessionId: "nonexist-session",
-			},
-		})
-
-		_, err := client.UpdateSessionParameters(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInvalidArgument, connectErr.Code())
-	})
-}
-
-func TestControllerService_UpdateUserRole(t *testing.T) {
-	t.Run("成功: ユーザーのロールを更新", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			UpdateUserRole(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.UpdateUserRoleResponse{
-				Role: "Admin",
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateUserRoleRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.UpdateUserRoleRequest{
-				SessionId: "session-123",
-				User:      &headlessv1.UpdateUserRoleRequest_UserId{UserId: "U-target"},
-				Role:      "Admin",
-			},
-		})
-
-		res, err := client.UpdateUserRole(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Equal(t, "Admin", res.Msg.GetRole())
-	})
-
-	t.Run("失敗: RPCクライアントの取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test2", "test2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test2", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateUserRoleRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.UpdateUserRoleRequest{
-				SessionId: "session-123",
-				User:      &headlessv1.UpdateUserRoleRequest_UserId{UserId: "U-target"},
-				Role:      "Admin",
-			},
-		})
-
-		_, err := client.UpdateUserRole(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_StartWorld(t *testing.T) {
-	t.Run("成功: ワールドを起動", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector - GetRpcClient
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		// Mock RPC call to start world
-		setup.mockRpcClient.EXPECT().
-			StartWorld(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.StartWorldResponse{
-				OpenedSession: &headlessv1.Session{
-					Id:   "session-new",
-					Name: "TestWorld",
-				},
-			}, nil)
-
-		worldUrl := "resrec:///U-test/R-12345"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.StartWorldRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.WorldStartupParameters{
-				LoadWorld: &headlessv1.WorldStartupParameters_LoadWorldUrl{LoadWorldUrl: worldUrl},
-			},
-		})
-
-		res, err := client.StartWorld(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.NotNil(t, res.Msg.GetOpenedSession())
-		assert.Equal(t, "session-new", res.Msg.GetOpenedSession().GetId())
-	})
-
-	t.Run("失敗: 存在しないホスト", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		worldUrl := "resrec:///U-test/R-12345"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.StartWorldRequest{
-			HostId: "nonexist-host",
-			Parameters: &headlessv1.WorldStartupParameters{
-				LoadWorld: &headlessv1.WorldStartupParameters_LoadWorldUrl{LoadWorldUrl: worldUrl},
-			},
-		})
-
-		_, err := client.StartWorld(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeNotFound, connectErr.Code())
-	})
-
-	t.Run("成功: forcePortが未指定の場合、ポート範囲から自動割り当て", func(t *testing.T) {
-		// 環境変数を設定
-		t.Setenv("SESSION_PORT_MIN", "30000")
-		t.Setenv("SESSION_PORT_MAX", "30100")
-
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test-port", "test-port@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test-port", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector - GetRpcClient
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		// Mock RPC call to start world
-		// forcePortが範囲内（30000-30100）で設定されていることを検証
-		setup.mockRpcClient.EXPECT().
-			StartWorld(gomock.Any(), gomock.Any(), gomock.Any()).
-			DoAndReturn(func(ctx any, req *headlessv1.StartWorldRequest, opts ...any) (*headlessv1.StartWorldResponse, error) {
-				// forcePortが自動割り当てされていることを確認
-				assert.NotEqual(t, uint32(0), req.GetParameters().GetForcePort(), "ForcePort should be auto-assigned")
-				assert.GreaterOrEqual(t, req.GetParameters().GetForcePort(), uint32(30000), "ForcePort should be >= 30000")
-				assert.LessOrEqual(t, req.GetParameters().GetForcePort(), uint32(30100), "ForcePort should be <= 30100")
-
-				return &headlessv1.StartWorldResponse{
-					OpenedSession: &headlessv1.Session{
-						Id:   "session-new",
-						Name: "TestWorld",
-					},
-				}, nil
-			})
-
-		worldUrl := "resrec:///U-test-port/R-12345"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.StartWorldRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.WorldStartupParameters{
-				LoadWorld: &headlessv1.WorldStartupParameters_LoadWorldUrl{LoadWorldUrl: worldUrl},
-				// ForcePortは未指定（0）
-			},
-		})
-
-		res, err := client.StartWorld(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.NotNil(t, res.Msg.GetOpenedSession())
-		assert.Equal(t, "session-new", res.Msg.GetOpenedSession().GetId())
-	})
-
-	t.Run("成功: forcePortが明示的に指定された場合、その値を使用", func(t *testing.T) {
-		// 環境変数を設定
-		t.Setenv("SESSION_PORT_MIN", "30000")
-		t.Setenv("SESSION_PORT_MAX", "30100")
-
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test-port2", "test-port2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test-port2", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector - GetRpcClient
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		// Mock RPC call to start world
-		// 明示的に指定したforcePortがそのまま使用されることを検証
-		expectedPort := uint32(40000)
-
-		setup.mockRpcClient.EXPECT().
-			StartWorld(gomock.Any(), gomock.Any(), gomock.Any()).
-			DoAndReturn(func(ctx any, req *headlessv1.StartWorldRequest, opts ...any) (*headlessv1.StartWorldResponse, error) {
-				// 明示的に指定したポートが使われていることを確認
-				assert.Equal(t, expectedPort, req.GetParameters().GetForcePort(), "ForcePort should be the explicitly specified value")
-
-				return &headlessv1.StartWorldResponse{
-					OpenedSession: &headlessv1.Session{
-						Id:   "session-new",
-						Name: "TestWorld",
-					},
-				}, nil
-			})
-
-		worldUrl := "resrec:///U-test-port2/R-12345"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.StartWorldRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.WorldStartupParameters{
-				LoadWorld: &headlessv1.WorldStartupParameters_LoadWorldUrl{LoadWorldUrl: worldUrl},
-				ForcePort: expectedPort, // 明示的にポートを指定
-			},
-		})
-
-		res, err := client.StartWorld(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.NotNil(t, res.Msg.GetOpenedSession())
-		assert.Equal(t, "session-new", res.Msg.GetOpenedSession().GetId())
-	})
-
-	t.Run("成功: 環境変数が未設定の場合、forcePortは自動割り当てされない", func(t *testing.T) {
-		// 環境変数を明示的にクリア
-		t.Setenv("SESSION_PORT_MIN", "")
-		t.Setenv("SESSION_PORT_MAX", "")
-
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test-noenv", "test-noenv@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test-noenv", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector - GetRpcClient
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		// Mock RPC call to start world
-		// 環境変数が未設定の場合、forcePortは0のまま（自動割り当てされない）
-		setup.mockRpcClient.EXPECT().
-			StartWorld(gomock.Any(), gomock.Any(), gomock.Any()).
-			DoAndReturn(func(ctx any, req *headlessv1.StartWorldRequest, opts ...any) (*headlessv1.StartWorldResponse, error) {
-				// forcePortが自動割り当てされていないことを確認（0のまま）
-				assert.Equal(t, uint32(0), req.GetParameters().GetForcePort(), "ForcePort should not be auto-assigned when env vars are not set")
-
-				return &headlessv1.StartWorldResponse{
-					OpenedSession: &headlessv1.Session{
-						Id:   "session-new",
-						Name: "TestWorld",
-					},
-				}, nil
-			})
-
-		worldUrl := "resrec:///U-test-noenv/R-12345"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.StartWorldRequest{
-			HostId: host.ID,
-			Parameters: &headlessv1.WorldStartupParameters{
-				LoadWorld: &headlessv1.WorldStartupParameters_LoadWorldUrl{LoadWorldUrl: worldUrl},
-				// ForcePortは未指定（0）
-			},
-		})
-
-		res, err := client.StartWorld(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.NotNil(t, res.Msg.GetOpenedSession())
-		assert.Equal(t, "session-new", res.Msg.GetOpenedSession().GetId())
-	})
-}
-
-func TestControllerService_InviteUser(t *testing.T) {
-	t.Run("成功: ユーザーIDでユーザーを招待", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			InviteUser(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.InviteUserResponse{}, nil)
-
-		userId := "U-target"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.InviteUserRequest{
-			HostId:    host.ID,
-			SessionId: "session-123",
-			User:      &hdlctrlv1.InviteUserRequest_UserId{UserId: userId},
-		})
-
-		res, err := client.InviteUser(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("成功: ユーザー名でユーザーを招待", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test2", "test2@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test2", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		setup.mockRpcClient.EXPECT().
-			InviteUser(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.InviteUserResponse{}, nil)
-
-		userName := "TargetUser"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.InviteUserRequest{
-			HostId:    host.ID,
-			SessionId: "session-123",
-			User:      &hdlctrlv1.InviteUserRequest_UserName{UserName: userName},
-		})
-
-		res, err := client.InviteUser(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("失敗: RPCクライアントの取得に失敗", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test3", "test3@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test3", "TestHost3", entity.HeadlessHostStatus_RUNNING)
-
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(nil, connect.NewError(connect.CodeInternal, nil))
-
-		userId := "U-target"
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.InviteUserRequest{
-			HostId:    host.ID,
-			SessionId: "session-123",
-			User:      &hdlctrlv1.InviteUserRequest_UserId{UserId: userId},
-		})
-
-		_, err := client.InviteUser(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInternal, connectErr.Code())
-	})
-}
-
-func TestControllerService_StopSession(t *testing.T) {
-	t.Run("成功: セッションを停止", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account, host, and session
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-		session := testutil.CreateTestSession(t, setup.queries, host.ID, "TestSession", entity.SessionStatus_RUNNING)
-
-		// Mock HostConnector - GetRpcClient
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil).
-			AnyTimes()
-
-		// Mock RPC call to get session
-		setup.mockRpcClient.EXPECT().
-			GetSession(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetSessionResponse{
-				Session: &headlessv1.Session{
-					Id:   session.ID,
-					Name: session.Name,
-				},
-			}, nil).
-			AnyTimes()
-
-		// Mock RPC call to stop session
-		setup.mockRpcClient.EXPECT().
-			StopSession(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.StopSessionResponse{}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.StopSessionRequest{
-			SessionId: session.ID,
-		})
-
-		res, err := client.StopSession(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("失敗: 存在しないセッション", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.StopSessionRequest{
-			SessionId: "nonexist-session",
-		})
-
-		_, err := client.StopSession(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeNotFound, connectErr.Code())
-	})
-}
-
-func TestControllerService_IssueResoniteLinkConnection(t *testing.T) {
-	t.Run("成功: 認証済みユーザがトークン付きURLを発行", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-		session := testutil.CreateTestSession(t, setup.queries, host.ID, "TestSession", entity.SessionStatus_RUNNING)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.IssueResoniteLinkConnectionRequest{
-			SessionId: session.ID,
-		})
-
-		res, err := client.IssueResoniteLinkConnection(t.Context(), req)
-		require.NoError(t, err)
-		require.NotNil(t, res.Msg)
-
-		// ws_path は path + query 形式
-		assert.True(t, strings.HasPrefix(res.Msg.GetWsPath(), "/resonite-link/ws?token="),
-			"ws_path should start with /resonite-link/ws?token=, got %q", res.Msg.GetWsPath())
-
-		// トークン部分を取り出して claims を検証
-		const prefix = "/resonite-link/ws?token="
-
-		token, decodeErr := url.QueryUnescape(strings.TrimPrefix(res.Msg.GetWsPath(), prefix))
-		require.NoError(t, decodeErr)
-
-		claims, err := auth.ParseResoniteLinkToken(token)
-		require.NoError(t, err)
-		assert.Equal(t, session.ID, claims.SessionID)
-		assert.NotEmpty(t, claims.UserID)
-		require.NotNil(t, res.Msg.GetExpiresAt())
-		assert.True(t, res.Msg.GetExpiresAt().AsTime().After(time.Now()))
-	})
-
-	t.Run("失敗: 認証なし", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// No Authorization header
-		req := connect.NewRequest(&hdlctrlv1.IssueResoniteLinkConnectionRequest{
-			SessionId: "any-session",
-		})
-
-		_, err := client.IssueResoniteLinkConnection(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeUnauthenticated, connectErr.Code())
-	})
-
-	t.Run("失敗: session_id が空", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.IssueResoniteLinkConnectionRequest{
-			SessionId: "",
-		})
-
-		_, err := client.IssueResoniteLinkConnection(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInvalidArgument, connectErr.Code())
-	})
-
-	t.Run("失敗: 存在しないセッション", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.IssueResoniteLinkConnectionRequest{
-			SessionId: "nonexist-session",
-		})
-
-		_, err := client.IssueResoniteLinkConnection(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeNotFound, connectErr.Code())
-	})
-}
-
-func TestControllerService_SearchSessions(t *testing.T) {
-	t.Run("成功: セッションを検索 (ページング検証)", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account, host, and 10 sessions for pagination
-		const totalSessions = 10
-
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		for i := 1; i <= totalSessions; i++ {
-			testutil.CreateTestSession(t, setup.queries, host.ID, fmt.Sprintf("Session%02d", i), entity.SessionStatus_RUNNING)
-		}
-
-		// Mock HostConnector - GetRpcClient (called when fetching host info)
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil).
-			AnyTimes()
-
-		// Mock RPC calls
-		setup.mockRpcClient.EXPECT().
-			GetAbout(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetAboutResponse{}, nil).
-			AnyTimes()
-
-		setup.mockRpcClient.EXPECT().
-			GetAccountInfo(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetAccountInfoResponse{}, nil).
-			AnyTimes()
-
-		setup.mockRpcClient.EXPECT().
-			GetStatus(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetStatusResponse{}, nil).
-			AnyTimes()
-
-		setup.mockRpcClient.EXPECT().
-			GetStartupConfigToRestore(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetStartupConfigToRestoreResponse{
-				StartupConfig: &headlessv1.StartupConfig{},
-			}, nil).
-			AnyTimes()
-
-		// Mock RPC call to list sessions (called by getHostSessions)
-		setup.mockRpcClient.EXPECT().
-			ListSessions(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.ListSessionsResponse{
-				Sessions: []*headlessv1.Session{},
-			}, nil).
-			AnyTimes()
-
-		// page 未指定 -> デフォルト 20 件 (10件しかないので全件)
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SearchSessionsRequest{
-			Parameters: &hdlctrlv1.SearchSessionsRequest_SearchParameters{},
-		})
-		res, err := client.SearchSessions(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-		assert.Len(t, res.Msg.GetSessions(), totalSessions)
-		require.NotNil(t, res.Msg.GetPage())
-		assert.Equal(t, int32(totalSessions), res.Msg.GetPage().GetTotalCount())
-		assert.Equal(t, int32(20), res.Msg.GetPage().GetPageSize())
-
-		// page_size=4 で 3 ページに分割
-		req2 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SearchSessionsRequest{
-			Parameters: &hdlctrlv1.SearchSessionsRequest_SearchParameters{},
-			Page:       &hdlctrlv1.PageRequest{PageIndex: 0, PageSize: 4},
-		})
-		res2, err := client.SearchSessions(t.Context(), req2)
-		require.NoError(t, err)
-		assert.Len(t, res2.Msg.GetSessions(), 4)
-		assert.Equal(t, int32(totalSessions), res2.Msg.GetPage().GetTotalCount())
-
-		req3 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SearchSessionsRequest{
-			Parameters: &hdlctrlv1.SearchSessionsRequest_SearchParameters{},
-			Page:       &hdlctrlv1.PageRequest{PageIndex: 1, PageSize: 4},
-		})
-		res3, err := client.SearchSessions(t.Context(), req3)
-		require.NoError(t, err)
-		assert.Len(t, res3.Msg.GetSessions(), 4)
-
-		req4 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SearchSessionsRequest{
-			Parameters: &hdlctrlv1.SearchSessionsRequest_SearchParameters{},
-			Page:       &hdlctrlv1.PageRequest{PageIndex: 2, PageSize: 4},
-		})
-		res4, err := client.SearchSessions(t.Context(), req4)
-		require.NoError(t, err)
-		assert.Len(t, res4.Msg.GetSessions(), 2)
-		assert.Equal(t, int32(totalSessions), res4.Msg.GetPage().GetTotalCount())
-	})
-
-	t.Run("失敗: 無効な検索パラメータ", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SearchSessionsRequest{
-			Parameters: &hdlctrlv1.SearchSessionsRequest_SearchParameters{
-				// Invalid parameters will be handled by usecase
-			},
-		})
-
-		_, err := client.SearchSessions(t.Context(), req)
-		// This may or may not fail depending on usecase implementation
-		// For now, just verify it returns a response or error
-		if err != nil {
-			connectErr := &connect.Error{}
-			ok := errors.As(err, &connectErr)
-			require.True(t, ok, "expected connect.Error")
-			assert.Equal(t, connect.CodeInvalidArgument, connectErr.Code())
-		}
-	})
-
-	t.Run("成功: StatusとHostIDでAND検索", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account and two hosts
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host1 := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost1", entity.HeadlessHostStatus_RUNNING)
-		host2 := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost2", entity.HeadlessHostStatus_RUNNING)
-
-		// Create sessions with different statuses on different hosts
-		testutil.CreateTestSession(t, setup.queries, host1.ID, "Host1-Ended1", entity.SessionStatus_ENDED)
-		testutil.CreateTestSession(t, setup.queries, host1.ID, "Host1-Ended2", entity.SessionStatus_ENDED)
-		testutil.CreateTestSession(t, setup.queries, host1.ID, "Host1-Running", entity.SessionStatus_RUNNING)
-		testutil.CreateTestSession(t, setup.queries, host2.ID, "Host2-Ended", entity.SessionStatus_ENDED)
-		testutil.CreateTestSession(t, setup.queries, host2.ID, "Host2-Running", entity.SessionStatus_RUNNING)
-
-		// Search for ENDED sessions on Host1 only
-		status := hdlctrlv1.SessionStatus_SESSION_STATUS_ENDED
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SearchSessionsRequest{
-			Parameters: &hdlctrlv1.SearchSessionsRequest_SearchParameters{
-				Status: &status,
-				HostId: &host1.ID,
-			},
-		})
-
-		res, err := client.SearchSessions(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-
-		// Should return only ENDED sessions from Host1 (2 sessions)
-		assert.Len(t, res.Msg.GetSessions(), 2)
-		require.NotNil(t, res.Msg.GetPage())
-		assert.Equal(t, int32(2), res.Msg.GetPage().GetTotalCount())
-
-		for _, session := range res.Msg.GetSessions() {
-			assert.Equal(t, host1.ID, session.GetHostId(), "Session should belong to Host1")
-			assert.Equal(t, hdlctrlv1.SessionStatus_SESSION_STATUS_ENDED, session.GetStatus(), "Session should be ENDED")
-		}
-
-		// 同条件で page_size=1 -> 1件ずつ取れる
-		req2 := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SearchSessionsRequest{
-			Parameters: &hdlctrlv1.SearchSessionsRequest_SearchParameters{
-				Status: &status,
-				HostId: &host1.ID,
-			},
-			Page: &hdlctrlv1.PageRequest{PageIndex: 0, PageSize: 1},
-		})
-		res2, err := client.SearchSessions(t.Context(), req2)
-		require.NoError(t, err)
-		assert.Len(t, res2.Msg.GetSessions(), 1)
-		assert.Equal(t, int32(2), res2.Msg.GetPage().GetTotalCount())
-	})
-}
-
-func TestControllerService_DeleteEndedSession(t *testing.T) {
-	t.Run("成功: 存在しないセッション（何も起こらない）", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.DeleteEndedSessionRequest{
-			SessionId: "nonexist-session",
-		})
-
-		_, err := client.DeleteEndedSession(t.Context(), req)
-		require.NoError(t, err)
-	})
-}
-
-func TestControllerService_UpdateSessionExtraSettings(t *testing.T) {
-	t.Run("成功: セッションの追加設定を更新", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test account, host, and session
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-test", "test@example.test", "password")
-		host := testutil.CreateTestHeadlessHost(t, setup.queries, "U-test", "TestHost", entity.HeadlessHostStatus_RUNNING)
-		session := testutil.CreateTestSession(t, setup.queries, host.ID, "TestSession", entity.SessionStatus_RUNNING)
-
-		// Mock HostConnector - GetRpcClient (called by GetSession)
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil).
-			AnyTimes()
-
-		// Mock RPC call to get session
-		setup.mockRpcClient.EXPECT().
-			GetSession(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetSessionResponse{
-				Session: &headlessv1.Session{
-					Id:   session.ID,
-					Name: session.Name,
-				},
-			}, nil).
-			AnyTimes()
-
-		autoUpgrade := true
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateSessionExtraSettingsRequest{
-			SessionId:   session.ID,
-			AutoUpgrade: &autoUpgrade,
-		})
-
-		res, err := client.UpdateSessionExtraSettings(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-
-		// Verify settings were updated in database
-		updatedSession, err := setup.queries.GetSession(t.Context(), session.ID)
-		require.NoError(t, err)
-		assert.True(t, updatedSession.AutoUpgrade)
-	})
-
-	t.Run("失敗: 存在しないセッション", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		autoUpgrade := true
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.UpdateSessionExtraSettingsRequest{
-			SessionId:   "nonexist-session",
-			AutoUpgrade: &autoUpgrade,
-		})
-
-		_, err := client.UpdateSessionExtraSettings(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeNotFound, connectErr.Code())
-	})
-}
-
-func TestControllerService_ListContacts(t *testing.T) {
-	t.Run("成功: コンタクト一覧を取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test headless account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-contact1", "contact@example.test", "password")
-		testutil.CreateTestHeadlessHost(t, setup.queries, "U-contact1", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector to return RPC client
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		// Mock RPC call to list contacts
-		setup.mockRpcClient.EXPECT().
-			ListContacts(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.ListContactsResponse{
-				Users: []*headlessv1.UserInfo{
-					{Id: "U-friend1", Name: "Friend1", IconUrl: "https://example.test/icon1.png"},
-					{Id: "U-friend2", Name: "Friend2", IconUrl: "https://example.test/icon2.png"},
-				},
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListContactsRequest{
-			HeadlessAccountId: "U-contact1",
-			Limit:             50,
-		})
-
-		res, err := client.ListContacts(t.Context(), req)
-		require.NoError(t, err)
-		assert.Len(t, res.Msg.GetContacts(), 2)
-		assert.Equal(t, "U-friend1", res.Msg.GetContacts()[0].GetId())
-		assert.Equal(t, "Friend1", res.Msg.GetContacts()[0].GetName())
-	})
-
-	t.Run("失敗: 起動中のホストがない", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test headless account without running host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-nohost", "nohost@example.test", "password")
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.ListContactsRequest{
-			HeadlessAccountId: "U-nohost",
-			Limit:             50,
-		})
-
-		_, err := client.ListContacts(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInvalidArgument, connectErr.Code())
-	})
-}
-
-func TestControllerService_GetContactMessages(t *testing.T) {
-	t.Run("成功: メッセージ履歴を取得", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test headless account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-msg1", "msg@example.test", "password")
-		testutil.CreateTestHeadlessHost(t, setup.queries, "U-msg1", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector to return RPC client
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		// Mock RPC call to get contact messages
-		setup.mockRpcClient.EXPECT().
-			GetContactMessages(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.GetContactMessagesResponse{
-				Messages: []*headlessv1.ContactChatMessage{
-					{Id: "msg1", Type: headlessv1.ContactChatMessageType_CONTACT_CHAT_MESSAGE_TYPE_TEXT, Content: "Hello", SenderId: "U-friend"},
-					{Id: "msg2", Type: headlessv1.ContactChatMessageType_CONTACT_CHAT_MESSAGE_TYPE_TEXT, Content: "Hi there", SenderId: "U-msg1"},
-				},
-				HasMoreBefore: true,
-				HasMoreAfter:  false,
-			}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetContactMessagesRequest{
-			HeadlessAccountId: "U-msg1",
-			ContactUserId:     "U-friend",
-			Limit:             50,
-		})
-
-		res, err := client.GetContactMessages(t.Context(), req)
-		require.NoError(t, err)
-		assert.Len(t, res.Msg.GetMessages(), 2)
-		assert.Equal(t, "Hello", res.Msg.GetMessages()[0].GetContent())
-		assert.False(t, res.Msg.GetMessages()[0].GetIsOwnMessage()) // Message from friend
-		assert.True(t, res.Msg.GetMessages()[1].GetIsOwnMessage())  // Message from self
-		assert.True(t, res.Msg.GetHasMoreBefore())
-		assert.False(t, res.Msg.GetHasMoreAfter())
-	})
-
-	t.Run("失敗: 起動中のホストがない", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test headless account without running host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-nomsghost", "nomsghost@example.test", "password")
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.GetContactMessagesRequest{
-			HeadlessAccountId: "U-nomsghost",
-			ContactUserId:     "U-friend",
-			Limit:             50,
-		})
-
-		_, err := client.GetContactMessages(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInvalidArgument, connectErr.Code())
-	})
-}
-
-func TestControllerService_SendContactMessage(t *testing.T) {
-	t.Run("成功: メッセージを送信", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test headless account and host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-send1", "send@example.test", "password")
-		testutil.CreateTestHeadlessHost(t, setup.queries, "U-send1", "TestHost", entity.HeadlessHostStatus_RUNNING)
-
-		// Mock HostConnector to return RPC client
-		setup.mockHostConnector.EXPECT().
-			GetRpcClient(gomock.Any(), gomock.Any()).
-			Return(setup.mockRpcClient, nil)
-
-		// Mock RPC call to send contact message
-		setup.mockRpcClient.EXPECT().
-			SendContactMessage(gomock.Any(), gomock.Any()).
-			Return(&headlessv1.SendContactMessageResponse{}, nil)
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SendContactMessageRequest{
-			HeadlessAccountId: "U-send1",
-			ContactUserId:     "U-friend",
-			Message:           "Hello, friend!",
-		})
-
-		res, err := client.SendContactMessage(t.Context(), req)
-		require.NoError(t, err)
-		assert.NotNil(t, res.Msg)
-	})
-
-	t.Run("失敗: 起動中のホストがない", func(t *testing.T) {
-		setup := setupControllerServiceTest(t)
-		defer setup.Cleanup()
-
-		client := setupAuthenticatedClient(t, setup.service)
-
-		// Create test headless account without running host
-		testutil.CreateTestHeadlessAccount(t, setup.queries, "U-nosendhost", "nosendhost@example.test", "password")
-
-		req := testutil.CreateDefaultAuthenticatedRequest(t, &hdlctrlv1.SendContactMessageRequest{
-			HeadlessAccountId: "U-nosendhost",
-			ContactUserId:     "U-friend",
-			Message:           "Hello!",
-		})
-
-		_, err := client.SendContactMessage(t.Context(), req)
-		require.Error(t, err)
-
-		connectErr := &connect.Error{}
-		ok := errors.As(err, &connectErr)
-		require.True(t, ok, "expected connect.Error")
-		assert.Equal(t, connect.CodeInvalidArgument, connectErr.Code())
 	})
 }
